### 2024-05-24 #6776 @editorialbot
[PRE REVIEW]: pide: Petrophysical Interpretation tools for geoDynamic Exploration.
**Submitting author:** @sinanozaydin (<a href="http://orcid.org/0000-0002-4532-9980">Sinan Özaydın</a>)
**Repository:** https://github.com/sinanozaydin/pide
**Branch with paper.md** (empty if default branch):
**Version:** v0.2
