### 2024-05-24 #6776 @kthyng
@sinanozaydin looks like the path to the figure in your paper isn't correct, can you fix this?
