### 2024-05-24 #6776 @kthyng
@sinanozaydin Comments:
* Docs: your notebooks look great! I would love to see them as hosted docs like readthedocs or similar if you get a chance (not a requirement)
* tests: I don't see a tests directory, though without looking in detail through your notebooks — can a user run through the notebooks themselves and verify that the software is working correctly for them by comparing output on their computer with the output in the notebooks? There needs to be a way to make sure the software is running correctly.
* We need to have the paper compiling correctly, see error above.
* Check DOI listings above and correct what you can
* Other items will be easier to verify once I can read your paper more easily.
