+++ docs/paper/paper.md
@@ -61,7 +61,7 @@ The `model` class, on the other hand, is where a collection of `material` object
 
 `pide` also comes with several tools that can exploit the program architecthure. 'model_modifier' functions, for instance, can be used to calculate certain specific hypothesis tests. `inversion` modules, on the other hand, can be utilised to invert for specific input parameters that fit real geophysical data.
 
-![Workflow Chart for pide \label{fig:pide_wflow}](/docs/figures/pide_workflow.png)
+![Workflow Chart for pide \label{fig:pide_wflow}](figures/pide_workflow.png)
 
 # Acknowledgements
 This study is supported by the Australian Research Council (ARC) Linkage Grant ARC-LP190100146 and ARC DP Grant ARC-DP220100709. 
