+++ docs/paper/paper.bib
@@ -6,6 +6,7 @@
   number={9},
   pages={e2020GC009126},
   year={2020},
+  doi={10.1029/2020gc009126},
   publisher={Wiley Online Library}
 }
 @article{Abers2016,
@@ -16,6 +17,7 @@
   number={2},
   pages={616--624},
   year={2016},
+  doi={10.1002/2015GC006171},
   publisher={Wiley Online Library}
 }
 @article{Pommier2011,
@@ -26,6 +28,7 @@
   number={9},
   pages={1450--1459},
   year={2011},
+  doi={10.1016/j.cageo.2011.01.002},
   publisher={Elsevier}
 }
 @article{Kelbert2014,
@@ -35,6 +38,7 @@
   volume={66},
   pages={40--53},
   year={2014},
+  doi={10.1016/j.cageo.2014.01.010},
   publisher={Elsevier}
 }
 @article{Key2016,
@@ -45,6 +49,7 @@
   number={1},
   pages={571--588},
   year={2016},
+  doi={10.1093/gji/ggw290},
   publisher={Oxford University Press}
 }
 @article{Soler2019,
@@ -55,5 +60,6 @@
   number={3},
   pages={2150--2164},
   year={2019},
+  doi={10.1093/gji/ggz277},
   publisher={Oxford University Press}
 }
