+++ docs/source/conf.py
@@ -1,0 +1,37 @@
+# Configuration file for the Sphinx documentation builder.
+#
+# For the full list of built-in configuration values, see the documentation:
+# https://www.sphinx-doc.org/en/master/usage/configuration.html
+
+# -- Project information -----------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
+
+import os
+import sys
+import sphinx_rtd_theme
+html_theme = 'sphinx_rtd_theme'
+
+
+sys.path.insert(0, os.path.abspath(".."))
+
+project = 'pide'
+copyright = '2024, Sinan Ozaydin, Lu Li, Utpal Singh, Patrice Rey, Maria Constanza Manassero'
+author = 'Sinan Ozaydin, Lu Li, Utpal Singh, Patrice Rey, Maria Constanza Manassero'
+release = '0.3'
+
+# -- General configuration ---------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
+
+extensions = ['sphinx.ext.autodoc','sphinx.ext.viewcode']
+
+templates_path = ['_templates']
+exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+
+
+# -- Options for HTML output -------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
+
+html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
+html_theme = 'sphinx_rtd_theme'
+html_static_path = ['_static']
+
