+++ pide/__init__.py
@@ -1,5 +1,5 @@
 """
-pide v.0.1.1
+pide v0.3
 
 """
 
@@ -7,8 +7,7 @@ import sys, csv, re, warnings, json
 
 from .pide import pide
 
-
 __author__ = 'Sinan Ozaydin'
-__version__ = '0.0.1'
+__version__ = '0.3'
 
 
