+++ docs/source/index.rst
@@ -1,0 +1,20 @@
+.. pide documentation master file, created by
+   sphinx-quickstart on Wed May 29 16:33:33 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to pide's documentation!
+================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
