+++ .readthedocs.yaml
@@ -4,11 +4,7 @@ version: 0.3
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.12"
-    # You can also specify other tool versions:
-    # nodejs: "19"
-    # rust: "1.64"
-    # golang: "1.19"
+    python: "3.9"
 
 sphinx:
   configuration: docs/source/conf.py
