### 2024-05-31 #6776 @sinanozaydin
@kthyng

Hello Kristen,

Thanks for your input! For readthedocs I am intending to build one in the near future. At this stage, is it enough to say it will be built later? My hands are really tied with fieldwork currently.

For tests, the users can use the notebooks for testing indeed.
