### 2024-06-24 #6776 @lucydot
Hi @kthyng ; I'm afraid this is far outside of my experience/training. It's probably best to wait for someone with a more suitable background to become free. However if there is no-one else in JOSS suited to it, then I will take it - it must be that sometimes submissions come up that are outside of everyones wheelhouse.
