### 2024-07-01 #6776 @kthyng
Hi @rowanc1 and @kujaku11! Are you interested in reviewing this submission for the Journal of Open Source Software? We ask that reviews start within a week or two and ideally wrap up within 4-6 weeks, keeping in mind this is usually an iterative process between the reviewers and author(s). We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Thanks for your consideration.
