### 2024-07-01 #6776 @kthyng
@kujaku11 Excellent! First I should have you read the conflict of interest policy here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy

If that is ok, then I'll add you as a reviewer. Once we have two reviewers I will start the review.
