### 2024-07-10 #6776 @kthyng
@rowanc1 I think you're at the same conference as me right now so you are busy now, but might you have bandwidth next week to take this review on?
