### 2024-07-13 #6776 @kthyng
Ok @rowanc1 made a good argument for not doing this review (not working in this field for years...) but @santisoler is also at SciPy and was willing to be pingex. @santisoler what do you think about reviewing this?
