### 2024-07-19 #6776 @santisoler
Hi @kthyng! I'm willing to review this submission.

I do have a question regarding the Conflict of Interest Policy. I see Lu Li as one of the contributors to the submitted article. Lu and I have been collaborating in open-source software development at @fatiando. We haven't collaborated in any publication besides an abstract submitted to AGU 2023 Fall Meeting ([more information here](https://github.com/santisoler/2023-agu)). Since Lu Li had not make significant contributions (3 commits vs 231 commits by @sinanozaydin), I consider there are not conflicts of interest for me to review this submission. I wanted to make it clear though so everyone is aware of this.

Hope this doesn't prevent me from reviewing this submission. I'll be waiting for your reply. Please, feel free to assign me as reviewer if you agree I'm able to do so.
