### 2024-07-22 #6776 @kthyng
@santisoler Thank you for declaring that information. Comparing it with our [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy), we are ok to move forward with you as a reviewer.
