### 2024-07-23 56f7d5c
missing parameters in material and model
