### 2024-07-23 dc43460
Adding the data files in the pypi (forgotten before)
