### 2024-07-23 68a159b
Adding the data files in the pypi (forgotten before)
