### 2024-07-23 d6fc43d
Adding the data files in the pypi (forgotten before)
