### 2024-07-24 193b514
updating to fix 3D interpolation and deformation related conductivity conversions
