### 2024-08-06 431bfce
post joss fixes to 3d conversion
