+++ pide/geodyn/deform_cond.py
@@ -5,17 +5,21 @@ from scipy.stats import linregress
 from scipy.optimize import curve_fit
 import warnings
 from scipy.optimize import OptimizeWarning
+import matplotlib.pyplot as plt
 
 warnings.simplefilter("ignore", OptimizeWarning)
 
 scipy_methods = ['lm', 'trf', 'dogbox']
 
-def check_misfit(x_array, y_array, x_vals, y_vals):
+def check_misfit(x_array, y_array, x_vals, y_vals, y_err):
 	
 	misfit_list = np.zeros(len(x_vals))
 	for i in range(0,len(x_vals)):
 		misfit_list[i] = y_array[(np.abs(x_array - x_vals[i])).argmin()] - y_vals[i]
-	return misfit_list
+	misfitsum = np.sum((misfit_list/y_err)**2.0)
+	rms = np.sqrt((1.0 / len(x_vals)) * misfitsum)
+
+	return rms, misfit_list
 
 def plastic_strain_2_conductivity(strain, low_cond, high_cond, low_strain, high_strain, function_method, **kwargs):
 	
@@ -35,6 +39,7 @@ def plastic_strain_2_conductivity(strain, low_cond, high_cond, low_strain, high_
 	
 		1. 'linear'
 		2. 'exponential'
+		3. 'polynomial'
 		
 	Optional parameters:
 		
@@ -46,26 +51,36 @@ def plastic_strain_2_conductivity(strain, low_cond, high_cond, low_strain, high_
 	return_all_params: Function returns all relevant parameters instead of just conductivity.
 	"""
 	
-	
-	if function_method == 'exponential':
-		cond_decay_modifier = -1
-	else:
-		cond_decay_modifier = 1
-		
-	conductivity_decay_factor = cond_decay_modifier * kwargs.pop('conductivity_decay_factor', 0.0)
-	strain_decay_factor = kwargs.pop('strain_decay_factor', 0.0)
+			
+	conductivity_decay_factor = kwargs.pop('conductivity_decay_factor', 0.7)
+	conductivity_decay_factor_2 = kwargs.pop('conductivity_decay_factor_2', 0.2)
+	strain_decay_factor = kwargs.pop('strain_decay_factor', 0.3)
 	return_all_params = kwargs.pop('return_all_params', False)
 	strain_percolation_threshold = kwargs.pop('strain_percolation_threshold', None)
+	index_plot = kwargs.pop('index_plot', 1)
 	
 	#Error values as kwargs:
-	low_cond_err = kwargs.pop('low_cond_err', 10)
-	mid_cond_err = kwargs.pop('mid_cond_err', 10)
-	high_cond_err = kwargs.pop('high_cond_err', 10)
+	cond_err = kwargs.pop('cond_err', 0.1)
 	
-	strain_func_build = np.logspace(np.log10(low_strain),np.log10(high_strain),100)
+	strain_func_build = np.log10(np.logspace(np.log10(low_strain),np.log10(high_strain),50))
 
-	def setup_data(low_strain, high_strain, low_cond, high_cond, strain_decay_factor, conductivity_decay_factor, low_cond_err = low_cond_err, 
-					high_cond_err = high_cond_err, mid_cond_err = mid_cond_err):
+	def setup_data(low_strain, high_strain, low_cond, high_cond, strain_decay_factor, conductivity_decay_factor, conductivity_decay_factor_2, cond_err = cond_err):
+					
+		if (strain_decay_factor < -0.9) or (strain_decay_factor >= 0.9):
+	
+			raise ValueError('The value entered for the strain_decay_factor is not valid. It has to be in between 0 to 0.9')
+		
[…]
