### 2024-08-06 ba7e849
post joss fixes to 3d conversion
