### 2024-08-11 sinanozaydin/pide#2 @kujaku11
JOSS Review: Example Notebook `10_2D_Underworld_conversion.ipynb`
Running Jupyter notebook from `examples/10_2D_Underworld_Conversion.ipynb` produces the following error in the 3rd cell from the bottom [using pip install --> v0.3 and notebooks from `main` branch]:

  The deformation related conductivity for the material  Sediment_0  is calculated.
  The deformation related conductivity for the material  Sediment_1  is calculated.
  The deformation related conductivity for the material  Sediment_1  is calculated.
  The deformation related conductivity for the material  Sediment_1  is calculated.
  The deformation related conductivity for the material  Sediment_1  is calculated.
  The deformation related conductivity for the material  Sediment_1  is calculated.
  The deformation related conductivity for the material  Sediment_1  is calculated.
RemoteTraceback                           Traceback (most recent call last)
RemoteTraceback:
"""
Traceback (most recent call last):
  File "\Anaconda3\envs\pide\lib\multiprocessing\pool.py", line 125, in worker
    result = (True, func(*args, **kwds))
  File "\Anaconda3\envs\pide\lib\multiprocessing\pool.py", line 48, in mapstar
    return list(map(*args))
  File "k\Anaconda3\envs\pide\lib\site-packages\pide\model.py", line 287, in run_deform2cond
    c, str_dcy, cnd_dcy, msft = plastic_strain_2_conductivity(strain = p_strain[index_number[0],index_number[1]],low_cond = background_cond[index_number[0],index_number[1]],
  File "\Anaconda3\envs\pide\lib\site-packages\pide\geodyn\deform_cond.py", line 184, in plastic_strain_2_conductivity
    cond_calced,strain_decay_factor,conductivity_decay_factor,misfit = fit(method = scipy_methods[0], init_start = True, run_idx = 0,
  File "\Anaconda3\envs\pide\lib\site-packages\pide\geodyn\deform_cond.py", line 127, in fit
    params, params_cov = curve_fit(func, strains, conds, method = method, sigma = cond_err)
  File "\Anaconda3\envs\pide\lib\site-packages\scipy\optimize\_minpack_py.py", line 925, in curve_fit
    ydata = np.asarray_chkfinite(ydata, float)
  File "\Anaconda3\envs\pide\lib\site-packages\numpy\lib\function_base.py", line 630, in asarray_chkfinite
    raise ValueError(
ValueError: array must not contain infs or NaNs
```
