### 2024-08-11 sinanozaydin/pide#3 @kujaku11
JOSS Review: Installation
Using `pip install pide` on Python 3.9 produces an error when trying install `pycifrw`

```
    ERROR: Failed building wheel for pycifrw
    Successfully built santex diffpy.structure
    Failed to build pycifrw
    ERROR: Could not build wheels for pycifrw, which is required to install pyproject.toml-based projects
```
Had to `conda install pycifrw -c conda-forge`
