### 2024-08-11 #7021 @kujaku11
@kthyng I have conducted my review and left an [issue](https://github.com/sinanozaydin/pide/issues/7) in the `pide` package repository.

Overall the package could be great contribution to the geophysics community.  Though there are some key components missing from the package.  Namely, **tests** and **documentation**.  Other issues include installation failed and some of the example notebooks don't run.  Once these are implemented `pide` will be a strong contribution.

The manuscript was fined, just some minor grammatical suggestions and suggestion to add an example figure from outputs of the program.

Let me know if you need anything else or clarity.
