### 2024-08-15 874b1b9
fix to cpu defaulting
