### 2024-08-16 e61203f
Update model.py
fix to 2d allocation of conductivity
