### 2024-08-16 b814b01
fix to one of the plag functions exception
