+++ pide/mt/mt_model_conversion.py
@@ -147,8 +147,6 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, core_bounds = No
 	num_vert_bounds = kwargs.pop('num_vert_bounds', 9)
 	vert_bound_incr = kwargs.pop('vert_bound_incr', 2)
 	
-	from pide.geodyn.interpolate_fields import interpolate_3d_fields
-
 	x_mesh = mesh[0]
 	y_mesh = mesh[1]
 	z_mesh = mesh[2]
@@ -161,8 +159,9 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, core_bounds = No
 				rho[i][j] = conductivity_array[(i*slice_len)+j]
 			except IndexError:
 				raise IndexError('The mesh structure entered does not match the conductivity array. Be sure the entered format mesh = (x_mesh_centers,y_mesh_centers,z_mesh_centers) in tuples are correct.')
-							
-	#determining the 
+					
+	
+	#determining the indexes
 	start_index_list = []
 	sea_index_list = []
 	for i in range(0,len(rho)):
@@ -171,7 +170,6 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, core_bounds = No
 		if (-999.0 in rho[i]) == False:
 			sea_index_list.append(i)
 	
-	air_start_index = start_index_list[0]
 	sea_index = sea_index_list[-1]
 	
 	for i in range(0,len(rho)):
@@ -206,9 +204,7 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, core_bounds = No
 	x_out = np.concatenate([xy_out[::-1],x_core,xy_out])
 	y_out = np.concatenate([yx_out[::-1],y_core,yx_out])
 	z_out = np.concatenate([z_core,z_out])
-	
-	n_len = len(x_out) * len(y_out)
-		
+			
 	rho_new = []
 	
 	for i in range(0,len(z_mesh)):
@@ -227,19 +223,19 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, core_bounds = No
 		rho_new.append(np.array(rho_local))
 	
 	rho_new = np.array(rho_new)
-
+	
 	for i in range(0,num_vert_bounds):
 		rho_new = np.insert(rho_new,0,np.ones(len(rho_new[-1])) * np.nan, axis = 0)
-		
+	
 	for i in range(0,len(rho)):
 		rho_new[i][0] = rho[i][0]
 		rho_new[i][len(x_core) + 2*len(xy_out)] = rho[i][len(x_core)]
 		rho_new[i][-(len(x_core) + 2*len(xy_out))] = rho[i][-len(x_core)]
 		rho_new[i][-1] = rho[i][-1]
 		
-	for i in range(len(rho),len(rho_new)):
-		rho_new[i] = rho_new[len(rho)]
-	
+	for i in range(0,num_vert_bounds):
+		rho_new[i] = rho_new[num_vert_bounds]
+		
 	rho_interp_array = rho_new.ravel()
 	
 	xi = np.cumsum(x_out)
@@ -254,7 +250,7 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, core_bounds = No
 	yi_n = [((yi[i] - yi[i-1]) / 2.0) + yi[i-1] for i in range(1, len(yi))]
 	zi_n = [((zi[i] - zi[i-1]) / 2.0) + zi[i-1] for i in range(1, len(zi))]
 	
-	zi_n = zi_n + z_mesh[-1] #adhjusting the air layers
+	zi_n = zi_n + z_mesh[-1] #adjusting the air layers
 	
 	zi_n = zi_n[::-1]
 	
[…]
