### 2024-08-21 b98a391
small fix to density calculation, if only a single mineral is calculated
