### 2024-08-22 sinanozaydin/pide#9 @santisoler
Missing documentation
> [!NOTE]
> This issue was opened as part of the JOSS review process: https://github.com/openjournals/joss-reviews/issues/7021

## The issue

Currently `pide` lacks a proper documentation, which is required by JOSS in order to accept the paper for publication.

The notebooks that showcase examples and tutorials on how to use the library are very welcomed when trying to familiarize with it and getting started. Nonetheless, these are not full replacements for documentation pages with API reference, installation instructions, community guidelines, etc (see [JOSS requirements for documentation](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation)).

## Possible solution

I would suggest the authors to fill the Sphinx skeleton they included in the `doc` folder with proper documentation pages for `pide`. These should follow JOSS requirements for documentation pages, including:

- Description of the library with a statement of need.
- Installation instructions.
- Instructions on how to use the different tools in the library.
- API reference, including a comprehensive list of all public functions and classes in the library, with their descriptions, list of arguments, attributes, and methods.
- Community guidelines on how to contribute to the software, how to report issues and how to seek support.

In order to write documentation pages, I recommend reading the [Writing documentation page of the Scientific Python Developer Guide](https://learn.scientific-python.org/development/guides/docs/).
