### 2024-08-22 sinanozaydin/pide#10 @santisoler
Missing test suite
> [!NOTE]
> This issue was opened as part of the JOSS review process: https://github.com/openjournals/joss-reviews/issues/7021

## The issue

One of the JOSS requirements in order to accept the article for publication is that the code repository includes a test suite that ensures that that the submitted software behaves as expected. Currently, `pide` doesn't include a test suite.

## Possible solution

I would suggest the authors to include a test suite that ensures that each unit of code in the repository works as expected. Achieving a 100% test coverage can be challenging, so I would suggest at least to include tests for the critical functions and classes of `pide`. Specially the ones that could lead to hard to detect misbehaviours and erroneous results that could significantly impact any scientific result obtained using `pide`.

After a test suite is created, I would also suggest to include instructions in the documentation on how to run the tests, and optionally set up a system to automatically run the tests (as [suggested by JOSS](https://joss.readthedocs.io/en/latest/review_criteria.html#tests)), like GitHub Actions.

In order to create a test suite, I would recommend taking a look at [`pytest` documentation](https://docs.pytest.org) and the [Test page of the Scientific Python Developer Guide](https://learn.scientific-python.org/development/guides/pytest/).
