### 2024-08-22 sinanozaydin/pide#11 @santisoler
Missing community guidelines
> [!NOTE]
> This issue was opened as part of the JOSS review process: https://github.com/openjournals/joss-reviews/issues/7021

## The issue

Currently `pide` lacks instructions on how to contribute to the repository and how to report issues. This is one of the JOSS requirements in order to accept the article for publication.

## Possible solution

I would suggest the authors to include community guidelines in the `README.md` and (optionally) in the documentation (see #9).
