### 2024-08-22 sinanozaydin/pide#12 @santisoler
Unused `pyproject.toml`
## The issue

Inside the `pide` folder there's one `pyproject.toml` file that is not being used while packaging `pide`. With the current layout, only the `setup.cfg` and `setup.py` files are being used.

## Possible solution

I would suggest the authors to either remove the unused `pyproject.toml`, or replace the `setup.cfg` and `setup.py` files for a single `pyproject.toml` to package `pide`.

> [!NOTE]
> This issue was opened as part of the JOSS review process: https://github.com/openjournals/joss-reviews/issues/7021

> [!IMPORTANT]
> This issue won't block the acceptance of the submitted article to JOSS, although I wanted to make the authors aware of it.
