### 2024-08-22 sinanozaydin/pide#14 @santisoler
PEP8 compliance for classes and functions names
> [!NOTE]
> This issue was opened as part of the JOSS review process: https://github.com/openjournals/joss-reviews/issues/7021

> [!IMPORTANT]
> This is a subjective recommendation, so feel free to ignore it if you disagree with it. Not addressing this comment won't block the acceptance of the submitted article to JOSS.

## The issue

Currently, `pide` is not using a [PEP8 naming convention](https://peps.python.org/pep-0008/#naming-conventions) for most of its classes and functions. For example, the `pide` class is defined in lowercase, while most of the functions to compute the different conductivity models use a mixture of capitalize with underscores (like `Liu2021_WetAlmandineGarnet_xFe()`).

PEP8 recommends using the camel case convention for class names (capitalize first letter of each word, don't separate with underscores (like `sklearn.linearl_model.LinearRegression`), and snake case convention for functions and methods (like `set_depth`).

## Possible solution

I think `pide` will benefit from applying PEP8 to the most relevant classes and functions, like renaming `pide.pide` to `pide.Pide`.
