### 2024-08-22 #7021 @santisoler
## Summary of first pass of my review

I think there's significant amount of work in `pide` that would make it a very useful tool for geophysical modelling and interpretation. I really appreciate the authors for the time they put into this project and that they made it available under an open-source license.
Currently, there are a few minimum requirements that the submitted article doesn't meet in order to be accepted for publication by JOSS. My major comments are related to missing documentation for users and contributors to understand how `pide` is designed, find out what tools are available, and missing tests to ensure that the code behaves as expected and reproduce the tests results.

I provide more details about these and other topics that should be address through issues in the upstream repository.

Looking forward to continue the conversation and the review process.

Thanks again to authors for the effort put into this project, and thanks @kthyng for the invitation to review it.
