### 2024-08-23 47a8903
elevation get function
