### 2024-08-26 a19db10
small detail on get elev
