### 2024-08-26 sinanozaydin/pide#3 @sinanozaydin
Hello @kujaku11,

I cannot recreate this error, or able to found any dependency for 'pycifrw' in any of the packages.
