+++ CONTRIBUTING.md
@@ -1,0 +1,21 @@
+# Contributing Guidelines for pide
+<img src="./docs/figures/pide_logo.png">
+
+## How to Contribute
+
+### Reporting Bugs
+
+If you find a bug, please file an issue on the repository with detailed information on how to reproduce the issue.
+
+### Proposing Features
+
+If you'd like to propose a new feature, please open an issue and start a discussion. Once the community and maintainers approve, you can submit a pull request.
+
+### Submitting Pull Requests
+
+- Fork the repository.
+- Create a new branch: `git checkout -b feature/my-new-feature`.
+- Make your changes.
+- Commit your changes: `git commit -m 'Add new feature'`.
+- Push to the branch: `git push origin feature/my-new-feature`.
+- Open a pull request.
