+++ pide/pyproject.toml
@@ -1,22 +1,0 @@
-[project]
-name = "pide"
-version = "0.3.0"
-authors = [
-  { name="Sinan Ozaydin", email="sinan.ozaydin@protonmail.com" },
-  { name="Lu Li"},
-  { name="Utpal Singh"},
-  { name="Patrice F. Rey"},
-  { name="Maria Constanza Manassero"}
-]
-description = "A python library for petrophysical calculations. "
-readme = "../README.md"
-requires-python = ">=3.8"
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: GPL-3.0",
-    "Operating System :: OS Independent",
-]
-
-[project.urls]
-Homepage = "https://github.com/sinanozaydin/pide"
-Issues = "https://github.com/sinanozaydin/pide/issues"
