### 2024-08-26 6b3e52f
first fixes to paper (@santisoler)
