+++ docs/paper/paper.md
@@ -54,12 +54,12 @@ The general workflow diagram of the library can ben seen in Figure 1. The librar
 
 `pide` is the main class in which the electrical conductivity and seismic-related observables are calculated. In order to achieve this, the relevant parameters have to be defined in the pide object with the associated functions (e.g., composition, water content, interconnectivity). Just using the `pide` class, for instance, the user can make a figure of all calculations of all the olivine electrical conductivity and seismic velocity models for olivine. While experimental parameters for electrical conductivity parameters are defined and calculated within the `pide`, seismic velocities of given earth materials are calculated through `SAnTex` library automatically.
 
-`material` is the class that can be specified as a holder of pre-defined material properties. For instance, one can create a Lherzolite material by mixing specific modal proportions of ol, opx, cpx and gt, how these constituents are interconnected, or how water behave among them. This class allows the user to define some parameters tied to other parameters via user-defined functions. This allows the users to streamline some specific processes in a modular manner.
+`material` is the class that can be specified as a holder of pre-defined material properties. For instance, one can create a Lherzolite material by mixing specific modal proportions of olivine (ol), orthopyroxene (opx), clinopyroxene (cpx) and garnet (gt), how these constituents are interconnected, or how water behave among them. This class allows the user to define some parameters tied to other parameters via user-defined functions. This allows the users to streamline some specific processes in a modular manner.
 
 The `model` class, on the other hand, is where a collection of `material` objects can be appended with specific positions indexed in 3D space.
-`model` is also where the user can calculate the magnetic and gravitational anomalies solely since these observables are dependent on the position of the materials and assigned magnetic and density parameters only. `pide` can generate synthetic data for magnetic and gravitational anomalies utilising the `harmonica` library [@Soler2019]. However, it can only generate synthetic magnetotelluric and seismic velocity models since their forward modelling is not as easy to implement within the pide framework. Currently, pide can generate input files for commonly used magnetotelluric modelling algorithms `ModEM` [@Kelbert2014] and `Mare2DEM` [@Key2016], and then users can generate synthetic data using the algorithms themselves with the given synthetic model.
+`model` is also where the user can calculate the magnetic and gravitational anomalies solely since these observables are dependent on the position of the materials and assigned magnetic and density parameters only. `pide` can generate synthetic data for magnetic and gravitational anomalies utilising the `harmonica` library [@fatiando_a_terra_project_2024]. *pide* can generate synthetic electric conductivity and seismic velocity models that can be saved as input files for commonly used magnetotelluric modelling algorithms `ModEM` [@Kelbert2014] and `Mare2DEM` [@Key2016]. Users then can generate synthetic data using the algorithms provided by these software packages.
 
-`pide` also comes with several tools that can exploit the program architecthure. 'model_modifier' functions, for instance, can be used to calculate certain specific hypothesis tests. `inversion` modules, on the other hand, can be utilised to invert for specific input parameters that fit real geophysical data.
+`pide` also comes with several tools that can exploit the program architecture. 'model_modifier' functions, for instance, can be used to calculate certain specific hypothesis tests. `inversion` modules, on the other hand, can be utilised to invert for specific input parameters that fit real geophysical data.
 
 ![Workflow Chart for pide \label{fig:pide_wflow}](figures/pide_workflow.png)
 
