+++ docs/paper/paper.bib
@@ -52,14 +52,25 @@
   doi={10.1093/gji/ggw290},
   publisher={Oxford University Press}
 }
-@article{Soler2019,
-  title={Gravitational field calculation in spherical coordinates using variable densities in depth},
-  author={Soler, Santiago R and Pesce, Agustina and Gimenez, Mario E and Uieda, Leonardo},
-  journal={Geophysical Journal International},
-  volume={218},
-  number={3},
-  pages={2150--2164},
-  year={2019},
-  doi={10.1093/gji/ggz277},
-  publisher={Oxford University Press}
+@software{fatiando_a_terra_project_2024,
+  author       = {Fatiando a Terra Project and
+                  Castro, Yago M. and
+                  Esteban, Federico D. and
+                  Li, Lu and
+                  Oliveira Jr, Vanderlei C and
+                  Pesce, Agustina and
+                  Shea, Nicholas and
+                  Soler, Santiago R. and
+                  Souza-Junior, Gelson F. and
+                  Tankersley, Matthew and
+                  Uieda, Leonardo and
+                  Uppal, India},
+  title        = {{Harmonica v0.7.0: Forward modeling, inversion, and 
+                   processing gravity and magnetic data}},
+  month        = aug,
+  year         = 2024,
+  publisher    = {Zenodo},
+  version      = {0.7.0},
+  doi          = {10.5281/zenodo.13308312},
+  url          = {https://doi.org/10.5281/zenodo.13308312}
 }
