### 2024-08-28 8b79da9
little fix to how to handle dimensions in model
