### 2024-09-03 b0ea557
fixes to phlogopite raise error and notebook 6
