### 2024-09-04 8c8dc53
added netCDF4 to required libraries.
