+++ pide/geodyn/surface_cond.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+
+import numpy as np
+
+def surface_cond()
