### 2024-09-04 5a6f239
changes for kujaku's comments
Changes to paper
changes to reference list
listing minerals now also prints out mechanisms
some fixes to notebooks
