### 2024-10-01 2cd7348
santex import update
santex has to be imported as:

from santex.isotropy import Isotropy
