### 2024-10-01 sinanozaydin/pide#12 @sinanozaydin
unused `pyproject.toml` is removed from the folder in the new version.
