### 2024-10-01 sinanozaydin/pide#11 @sinanozaydin
Contributing guidlines is added to the CONTRIBUTING.md, which has a link from README.md.
