### 2024-10-01 sinanozaydin/pide#2 @sinanozaydin
The issues should be resolved in the latest version at the github repo (post-joss branch).
