### 2024-10-01 10d3abd
some fixes and started testing
