### 2024-10-02 383808d
some fixes to elcond library
