### 2024-10-02 bc21291
test for experimental electrical conductivity measurements
