+++ tests/test_electrical_cond.py
@@ -1,23 +1,35 @@
 import pide
 import unittest
-
-
+import numpy as np
 
 class test_conductivity_functions(unittest.TestCase):
 
 	@classmethod
 	def setUpClass(self):
+	
+		#Test written for electrical conductivity models at October 2 2024. With newly added functions, this test file has to change
+		#since there are new studies added to the arrays.
 		
-		temp = [1000]
-		pres = 2.0
+		#tolerance of acceptance value
+		self.atol = 1e-5
+		
+		#Setting up temperature and pressure
+		temp = np.array([1000])
+		pres = np.array(2.0)
+		
 
-		p_obj = pide.pide()
-
-		p_obj.set_temperature(temp)
-		p_obj.set_pressure(pres)
-		p_obj.set_param1_mineral(mica = 0.2, plag = 0.1)
-
-		mineral_cond_list_check = [[0.0002147666215751601, 0.0019628863936549724, 0.0019628863936549724, 1.0190583930014192e-05,
+		self.p_obj = pide.pide()
+		
+		#setting up temperature pressure and equivalent parameters used in calculations
+		self.p_obj.set_temperature(temp)
+		self.p_obj.set_pressure(pres)
+		self.p_obj.set_param1_mineral(mica = 0.2, plag = 0.1)
+		self.p_obj.set_param1_rock(sandstone = 10)
+		self.p_obj.set_fluid_properties(salinity = 0.1)
+		self.p_obj.set_melt_properties(co2 = 100, water = 1000, na2o = 0.1, k2o = 0.1)
+		
+		#calculated values for all minerals and rocks
+		self.mineral_cond_list_check = [[0.0002147666215751601, 0.0019628863936549724, 0.0019628863936549724, 1.0190583930014192e-05,
 						8.208132334583077e-05, 0.00020788083347927274, 0.0001440326145468413, 6.855909754207005e-07, 0.0010714138960102133,
 						7.377942552234648e-06, 1.7927554236795322e-06, 3.297803999685759e-06, 1.7484794023483996e-06,
 						0.00018056600809558101, 1.6910366224498637e-06, 1.325877798709181e-05, 5.684421480001739e-06, 3.826141179766439e-06,
@@ -50,8 +62,99 @@ class test_conductivity_functions(unittest.TestCase):
 						0.0005317099241385271, 0.0025914554862907784, 0.04564406736998614, 0.14886276382806088,
 						0.7179351351044376, 2.4126573631693167], [244.21164927494843, 0.0027744528159345428, 0.003604655509112567,
 						0.0003936708988150554]]
+						
+		self.rock_cond_list_check = [[6.850646712231316e-06, 4.289840479223029e-05, 0.01770295328980004, 0.0002448778106938003, 0.00021047417616575027, 0.00012485445378909127,
+		4.200470588638835e-06, 0.00033903506939415547, 0.0003958561083780382, 8.315216946910493e-05, 1.2486261946646642e-05, 0.00024114209299575372],
+		[0.024589951680813342, 3.63368464619912, 0.0036273508031219375, 0.0069424914042069885, 0.04346239059278374, 0.006664214605720846], 
+		[0.031188895840939354], [0.002073930843385905, 0.00745376397205312, 0.01218077065864558, 0.0775612069832346, 0.46737810548631803],
+		[0.14736974502790565, 3.9464210219680207, 58.54288802397429, 13667.540326691233, 0.008290972861156798], [0.0012790228002886473, 
+		0.06626895617758016, 0.012726584640203181, 0.004621835463107331, 0.0013870339125108635, 0.003985402950653599, 0.15773933062288584], 
[…]
