### 2024-10-02 sinanozaydin/pide#14 @sinanozaydin
I was not aware of these standards before. Since this is a subjective recommendation and I do not really have much time, I decided not to go forward with this suggestion. Thank you!
