### 2024-10-02 260b1d4
some fixes to pide and new tests for water distribution and solubility
