### 2024-10-03 4e7f0ae
some fixes for pide and tests for pide.py calculation functions
