### 2024-10-03 a99c6e2
tests for inversion, material and model
