+++ tests/inversion/test_inversion_geotherm.py
@@ -1,0 +1,150 @@
+import pide
+import unittest
+import numpy as np
+
+class test_inversion_and_geotherm(unittest.TestCase):
+
+	@classmethod
+	def setUpClass(self):
+	
+		#Test written for electrical conductivity models at October 2 2024. With newly added functions, this test file has to change
+		#since there are new studies added to the arrays.
+		
+		#tolerance of acceptance value
+		self.atol = 1e-3
+		
+		self.geotherm_check = [298.0, 310.4200028903184, 322.71369349550923, 334.9758311552669, 347.1877592106995, 359.3328353384015,
+		371.39543370915834, 383.36093207763616, 395.2156851840189, 406.9469936865668, 418.5430696791802, 429.99299974077053,
+		441.2867063149084, 452.4149080878951, 463.36907991964256, 474.1414127831915, 484.72477408387454, 495.11266865651777,
+		506.5156036964089, 517.766514458218, 528.8639956679107, 539.8047128767944, 550.5856255510686, 561.2039613809226,
+		571.6571963533186, 581.9430363933021, 592.0594004424398, 602.004404848924, 611.7763489555439, 621.3737017823543,
+		630.7950897105463, 640.0392850827975, 649.1051956433168, 657.9918547479825, 666.6984122814409, 675.224126223883,
+		683.5683548154775, 691.7305492711894, 699.7102470029878, 707.5070653103031, 712.7054253199909, 717.901615320773,
+		723.092194969475, 728.2770188547134, 733.4559403882583, 738.6288158725082, 743.7955044381787, 748.955867988141,
+		754.1097711422777, 759.25708118335, 764.3976680038658, 769.5314040539371, 774.5343649208719, 779.53007088,
+		784.5182830174514, 789.498894796334, 794.4718016473095, 799.4369010617172, 804.3940925495335, 809.3432775983132,
+		814.2843596329607, 819.2172439763139, 824.1418378105278, 829.0580501392437, 833.9657917505289, 838.864975180573,
+		843.7555146781286, 848.63732616968, 853.5103272253273, 858.3744370253736, 863.2295763275998, 868.0756674352158,
+		872.9126341654732, 877.7404018189281, 882.5588971493413, 887.3680483342016, 892.1677849458625, 896.9580379232782,
+		901.7387395443274, 906.5098233987138, 911.2712243614316, 916.0228785667828, 920.7647233829382, 925.4966973870288,
+		930.2187403407572, 934.930793166519, 939.6327979240233, 944.324697787402, 949.0064370227981, 953.677960966423,
+		958.3392160030731, 962.9901495450958, 967.6307100117965, 972.2608468092764, 976.8805103106928, 981.4896518369331, 
+		986.0882236376934, 990.6761788729531, 995.2534715948383, 999.820056729865, 1004.3758900615542, 1008.9209282134107,
+		1013.4551286322595, 1017.9784495719306, 1022.4908500772851, 1026.992289968577, 1031.4827298261418, 1035.962130975405,
+		1040.4304554722064, 1044.8876660884298, 1049.3337262979342, 1053.76860026278, 1058.1922528197422, 1062.6046494671075,
+		1067.005756351746, 1071.395540256454, 1075.7739685875629, 1080.1410093628047, 1084.496631199434, 1088.8408033025955,
[…]
