### 2024-10-03 sinanozaydin/pide#10 @sinanozaydin
Thank you for the reccomendations. Test suits is now built on the post_joss branch of pide. We included test functions for the following parts of the library:

- Main pide calculation functions such as electrical conductivity, water solubility and distribution.
- Model and Material classes.
- Inversion module.

We will include how to run these tests in the documentation shortly.
