### 2024-10-03 3decf18
Delete inversion_results.py
