### 2024-10-03 111ad4e
changes to docstrings pt1
