### 2024-10-04 b87ac60
changes to docstrings pt2
