+++ pide/utils/gis_tools.py
@@ -4,16 +4,30 @@ import numpy as np
 
 def get_utm_zone_number(longitude):
 
-	#A function to return the utm number dependent on the longitude
+	"""
+	A function to get utm zone number
+	Input:
+	float: longitude
+	
+	Output:
+	UTM Zone Number
+	"""
 
 	zone_number = math.ceil((longitude + 180.0) / 6.0)
 	
 	return zone_number
 
 def lat_lon_to_utm(latitude, longitude, zone_number = None):
-
-	#A function to convert latitude longitude coordinates in WGS84 projection to UTM coordinates.
-
+	"""
+	A function to convert latitude longitude coordinates in WGS84 projection to UTM coordinates.
+	
+	Input:
+	float or array: latitude
+	float or array: longitude
+	
+	Output:
+	UTM Coordinates
+	"""
 	wgs84 = Proj(proj='latlong', datum='WGS84')
 	
 	if zone_number == None:
@@ -30,9 +44,9 @@ def lat_lon_to_utm(latitude, longitude, zone_number = None):
 	return utm_x, utm_y
 	
 def utm_to_lat_lon(utm_x, utm_y, zone_number):
-	
-	# A function to convert UTM coordinates to latitude longitudes in WGS84 projectin.
-		
+	"""
+	A function to convert UTM coordinates to latitude longitudes in WGS84 projectin.
+	"""
 	utm = Proj(proj='utm', zone=zone_number, datum='WGS84', ellps='WGS84')
 
 	longitude, latitude = transform(utm, Proj(proj='latlong', datum='WGS84'), utm_x, utm_y)
