+++ pide/material.py
@@ -7,7 +7,7 @@ class Material(object):
 
 	def __init__(self, name = "Unnamed", material_index = None, calculation_type = 'mineral', composition = None, melt_fluid_frac = 0.0,
 	interconnectivities = None, param1 = None, el_cond_selections = None, melt_fluid_incorporation_method = 'Field', melt_or_fluid = 'melt', melt_fluid_m = 8.0,
-	melt_properties = None,	melt_fluid_cond_selection = None, water_distr = False, water = None, xfe = None, solid_phase_mixing_idx = 0, melt_fluid_phase_mixing_idx = 0,
+	melt_properties = None, fluid_properties = None, melt_fluid_cond_selection = None, water_distr = False, water = None, xfe = None, solid_phase_mixing_idx = 0, melt_fluid_phase_mixing_idx = 0,
 	deformation_dict = None, top = None, bottom = None, **kwargs):
 	
 		"""
@@ -39,6 +39,9 @@ class Material(object):
 		dict: melt_properties - properties of melt entered in property_name:value fashion. 
 											      'water','co2','na2o','k2o'
 											      
+		dict: fluid_properties - properties of fluid entered in property_name:value fashion.
+												   'salinity'
+											      
 		int: melt_fluid_cond_selection - electrical conductivity model chosen for melt/fluid phase.
 		
 		bool: water_distr - Boolean to automatically distribute water among the phases or not.
@@ -62,8 +65,19 @@ class Material(object):
 		
 		dict: mantle_water_part - mantle water partitioning dictionart entered in mineral_name:index fashion.
 		
+		float: resistivity_medium - resistivity of the material entered as a value instead of empirical calculation. || in ohm.m
 		
+		float: vp_medium - P-wave velocity value of the material entered as a value instead of empirical calculation. || in km/s
 		
+		float: vs_medium - S-wave velocity value of the material entered as a value instead of empirical calculation. || in km/s
+
+		dict: water_calib - dictionary to enter water calibration methods for water calculations entered in material:mode fashion.
+												'ol', 'px_gt', 'plag'
+		int: o2_buffer - selection of o2 buffer with denoted integer value.
+		
+		array: al_opx - Al in orthopyroxene in array form || in w.t.%
+		
+		float: magnetic_susceptibility - magnetic susceptibility of the material || in Am^-1
 		
 		
 		---------------------------------- -----------------------------------------
@@ -168,6 +182,11 @@ class Material(object):
 		self._melt_properties = None
 		self.melt_properties = melt_properties
 		
+		if fluid_properties == None:
+			fluid_properties = {'salinity': 0.1}
+		self._fluid_properties = None
+		self.fluid_properties = fluid_properties
+		
 		if deformation_dict == None:
 			deformation_dict = {'function_method':'linear','conductivity_decay_factor':0, 'conductivity_decay_factor_2':0,
 			'strain_decay_factor':0, 'strain_percolation_threshold': None}
@@ -177,9 +196,7 @@ class Material(object):
 				
 		self._mantle_water_part = None
 		self.mantle_water_part = kwargs.pop('mantle_water_part', {'opx_ol':0,'cpx_ol':0,'garnet_ol':0, 'ol_melt':0, 'opx_melt':0, 'cpx_melt':0,'garnet_melt':0})
-		
-		self.mantle_water_sol_ref = kwargs.pop('mantle_water_sol_ref', 'ol')
-		
+				
 		self.resistivity_medium = kwargs.pop('resistivity_medium', None)
 		self.vp_medium = kwargs.pop('vp_medium', None)
[…]
