+++ .readthedocs.yaml
@@ -1,4 +1,4 @@
-version: 0.3
+version: 2
 
 # Set the OS, Python version and other tools you might need
 build:
