### 2024-10-04 b600734
added requirements.txt
