+++ docs/requirements.txt
@@ -1,0 +1,4 @@
+sphinx
+sphinx-rtd-theme
+nbsphinx
+recommonmark
