+++ docs/requirements.txt
@@ -2,3 +2,4 @@ sphinx
 sphinx-rtd-theme
 nbsphinx
 recommonmark
+jupyter
