+++ docs/source/index.rst
@@ -10,8 +10,12 @@ Welcome to documentation for pide!
    :maxdepth: 2
    :caption: Contents:
 
-   ../examples/notebooks/0_pide_general_tutorial.ipynb
-   0_pide_general_tutorial.ipynb
+.. include:: ../../README.md
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Notebooks
+   ../../examples/notebooks/0_pide_general_tutorial.ipynb
 
 Indices and tables
 ==================
