+++ docs/source/index.rst
@@ -1,25 +1,0 @@
-.. pide_rtd documentation master file, created by
-   sphinx-quickstart on Fri Oct  4 16:10:09 2024.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to documentation for pide!
-====================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-.. include:: ../../README.md
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Notebooks
-   ../../examples/notebooks/0_pide_general_tutorial.ipynb
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
