+++ docs/requirements.txt
@@ -1,6 +1,5 @@
 sphinx>=4.0
 sphinx-rtd-theme
 nbsphinx
-recommonmark
 jupyter
 myst-parser
