+++ docs/source/index.md
@@ -1,23 +1,9 @@
-.. pide_rtd documentation master file, created by
-   sphinx-quickstart on Fri Oct  4 16:10:09 2024.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-
 ====================================
 
 # Welcome to documentation for pide!
 
+!INCLUDE "../../README.md"
+
 ## Contents
 
-```{toctree}
-:maxdepth: 2
-:caption: Contents:
 
-```{include} ../../README.md
-
-## Indices and tables
-
-* {ref}`genindex`
-* {ref}`modindex`
-* {ref}`search`
