+++ docs/source/index.md
@@ -3,7 +3,7 @@
 # Welcome to documentation for pide!
 
 # pide
-<img src="../figures/pide_logo.png">
+![Pide Logo](../figures/pide_logo.png)
 
 # About
 *pide* is a Python3 library for calculating geophysical parameters (e.g., electrical conductivity, seismic velocity), employing the results from experimental petrology, mineral/rock physics, and thermomechanical modelling studies. *pide* can calculate the theoretical electrical conductivity of any earth material that exists in the literature. *pide* can also calculate seismic velocity utilising the external 'sister' library *santex*. Using these theoretical calculations, users can utilise inversion modules to decode geophysical anomalies compositionally or convert thermomechanical models into geophysical observables. With a given spatial mapping of earth materials, which can preferentially be loaded from a thermomechanical model, *pide*  can be used to build synthetic electrical conductivity and seismic velocity models and generate gravity and magnetic anomalies. Moreover, it is built as a modular tool, so users can easily build their functions.
@@ -26,7 +26,7 @@ If you want to help develop and change code as it is being used:
 
 How to use *pide* can be learned through Jupyter notebooks provided in examples/notebooks directory. The general workflow can be tracked through the chart below:
 
-<img src="../paper/figures/pide_workflow.png">
+![Pide Logo](../paper/figures/pide_workflow.png)
 
 Information on all methods (input/output, examples) can be accesed through the pide object method as follows:
 <pre>
@@ -47,7 +47,7 @@ Since *pide* is an open-source library, users are encouraged to be contribute an
 | **Sinan Özaydın** | sinan.ozaydin@protonmail.com | sinan.ozaydin@sydney.edu.au
 
 # General Tutorial
-```{include} ../../examples/notebooks/0_pide_general_tutorial.ipynb
+```{include} 0_pide_general_tutorial.md
 
 
 
