+++ docs/source/8_MantleWaterContent_and_Rheology_from_MT.md
@@ -1,0 +1,283 @@
+## <span style="color:green"> Notebook - Mapping metasomatised domains from MT Models</span>
+<span style="color:purple">Sinan Ozaydin, School of Geosciences, The University of Sydney, NSW 2006, Australia <br/> </span>
+<span style="color:purple">sinan.ozaydin@protonmail.com | sinan.ozaydin@sydney.edu.au </span>
+
+
+```python
+import os
+from pathlib import Path
+import numpy as np
+from scipy.interpolate import griddata
+import matplotlib.pyplot as plt
+from netCDF4 import Dataset
+
+import pide
+from pide.mt.mt_model_read import read_ModEM_rho
+from pide.utils.utils import read_csv, associate_coordinates
+from pide.utils.gis_tools import lat_lon_to_utm, get_utm_zone_number
+from pide.inversion import conductivity_solver_single_param
+
+import time
+
+notebook_path = Path().resolve()
+```
+
+In this notebook, we will convert the MT models of Western United States into water content maps at 96 and 150 km. This will require a simple inversion process with the given environment. We will use the same cells in Notebook 7 to read the models and assign the relevant indexes.
+
+First of all, let's read the MT model by using the ***read_ModEM_rho*** function and determine the model center in UTM coordinates to georeference the model using the functions from ***pide.utils.gis_tools***. We will be using the depths according to 96 and 155 kilometer slices to calculate our water content maps.
+
+
+```python
+wus_rho_path = os.path.join(notebook_path,'..','example_data','WUS_MT_Model','WUS.MT.Bedrosian2021.resistivity.rho')
+
+#model centers for georeferencing
+mc_lat = 40.75
+mc_lon = -113.25
+
+utm_no = get_utm_zone_number(mc_lon)
+mc_x, mc_y = lat_lon_to_utm(mc_lat,mc_lon)
+
+#reading the MT model file
+rho, mesh_x, mesh_y, mesh_z = read_ModEM_rho(rho_file_path=wus_rho_path)
+
+#converting x and y to utm coordinates
+x = mesh_x
+y = mesh_y
+
+#Getting the vertical indexes of 
+idx_96 = (np.abs((mesh_z/1e3) - 96.0)).argmin()
+idx_155 = (np.abs((mesh_z/1e3) - 155.0)).argmin()
+
+#Getting the resistivities at 50 and 100 km
+rho_96 = rho[idx_96].flatten()
+rho_155 = rho[idx_155].flatten()
+
+cond_96 = 1.0/rho_96
+cond_155 = 1.0/rho_155
+```
+
+    /home/sinan/src/SEL/pide/utils/gis_tools.py:28: FutureWarning: This function is deprecated. See: https://pyproj4.github.io/pyproj/stable/gotchas.html#upgrading-to-pyproj-2-from-pyproj-1
+      utm_x, utm_y = transform(wgs84, utm, longitude, latitude)
+
+
+    [0.01186123 0.01439003 0.0154702  ... 0.00620215 0.00595479 0.00678921]
+    [0.00776647 0.0060221  0.00449069 ... 0.00668967 0.00642383 0.00718076]
+
+
+Now, let's read the temperature model with the help of utils.read_csv function. and then georefence them into the UTM coordinates.
+
+
+```python
+#Reading the temperature model
[…]
