+++ docs/source/2_PhaseMixing_and_WaterDistribution.md
@@ -1,0 +1,386 @@
+## <span style="color:green"> Notebook - Phase Mixing Functions and Water Distribution </span>
+<span style="color:purple">Sinan Ozaydin, School of Geosciences, The University of Sydney, NSW 2006, Australia <br/> </span>
+<span style="color:purple">sinan.ozaydin@protonmail.com | sinan.ozaydin@sydney.edu.au </span>
+
+This notebook will demonstrate how to use the phase mixing functions and water distribution works for electrical conductivity calculation. Phase mixing functions are used to calculate the bulk conductivity of a mineral assemblage with the given geometrical assumptions. Since electrical conductivity is a measure of how easily can electrons can flow through a system, the connectivity of conductive phases in a matrix has crucial importance when it comes to interpreting the electrical conductivity variations measured with MT method.
+
+In ***pide***, there are six solid-state phase mixing functions are available to use. Which can be listed with the code snippet below:
+
+
+
+```python
+import pide
+import numpy as np
+import matplotlib.pyplot as plt
+
+p_obj = pide.pide() #forming the pide object
+mixing_list = p_obj.list_phs_mix_methods()
+```
+
+    [91mSolid Phase Mixing Models:[0m
+    0.   Generalized Archie's Law (Glover, 2010)
+    1.   Hashin-Shtrikman Lower Bound (Berryman, 1995)
+    2.   Hashin-Shtrikman Upper Bound (Berryman, 1995)
+    3.   Parallel Model (Guegen and Palciauskas, 1994)
+    4.   Perpendicular Model (Guegen and Palciauskas, 1994)
+    5.   Random Model (Guegen and Palciauskas, 1994)
+
+
+    /home/sinan/.local/lib/python3.10/site-packages/matplotlib/projections/__init__.py:63: UserWarning: Unable to import Axes3D. This may be due to multiple versions of Matplotlib being installed (e.g. as a system package and as a pip package). As a result, the 3D projection is not available.
+      warnings.warn("Unable to import Axes3D. This may be due to multiple versions of "
+
+
+$$\sigma = \sum_{i = 1}^n \sigma_{i} \phi^{m} \qquad \text{(1)} \quad \text{Generalized Archie's Law}$$
+$$\sigma = \left( \sum_{i = 1}^n \frac{\phi_i}{\sigma_i + 2\sigma_{min}}   \right)^{-1} - 2\sigma_{min} \qquad \text{(2)} \quad \text{Hashin-Shtrikman Lower Bound}$$
+$$\sigma = \left( \sum_{i = 1}^n \frac{\phi_i}{\sigma_i + 2\sigma_{max}}   \right)^{-1} - 2\sigma_{max} \qquad \text{(3)} \quad \text{Hashin-Shtrikman Upper Bound}$$
+$$\sigma = \sum_{i = 1}^n \phi_i \sigma_i \quad \text{(4)} \quad \text{Parallel Model}$$
+$$\sigma = \sum_{i = 1}^n \frac{\sigma_i}{\phi_i} \quad \text{(5)} \quad \text{Perpendicular Model}$$
+$$\sigma = \prod \sigma_i^{\phi_i} \quad \text{(6)} \quad \text{Random Model}$$
+
+
+where volumetrically dominant cementation exponent for m in Generalized Archie's law is:
+
[…]
