+++ docs/source/3_Seismic_Velocity_Calculation.md
@@ -1,191 +1,0 @@
-## <span style="color:green"> Notebook - Calculating Seismic Velocities </span>
-<span style="color:purple">Sinan Ozaydin, School of Geosciences, The University of Sydney, NSW 2006, Australia <br/> </span>
-<span style="color:purple">sinan.ozaydin@protonmail.com</span>
-
-In this notebook, we we will explore how to calculate seismic velocities using the library pide for a given earth material. Seismic velocities in **pide** library are calculated by utilising the 'sister' library of pide **santex**. The isotropic velocities are calculated by the methods described in **Hacker et al., (2003)**. Material elastic parameters entered for all minerals are stored in materials.json file in pide_src and isotropy folder in santex. When a rock entered for the composition in pide, associated mineral assemblages in correspondent electrical conductivity models chosen are automatically calculated. Bulk velocities of the material formed are calculated by using Hashin-Shtrikman mixing function.
-
-
-```python
-import pide
-import numpy as np
-import matplotlib.pyplot as plt
-```
-
-
-    ---------------------------------------------------------------------------
-
-    ImportError                               Traceback (most recent call last)
-
-    Cell In[1], line 1
-    ----> 1 import pide
-          2 import numpy as np
-          3 import matplotlib.pyplot as plt
-
-
-    File ~/src/SEL/pide/__init__.py:8
-          1 """
-          2 pide v0.3
-          3 
-          4 """
-          6 import sys, csv, re, warnings, json
-    ----> 8 from .pide import pide
-         10 __author__ = 'Sinan Ozaydin'
-         11 __version__ = '0.3'
-
-
-    File ~/src/SEL/pide/pide.py:10
-          8 import numpy as np
-          9 from scipy.interpolate import interp1d
-    ---> 10 from santex import Isotropy
-         12 sys.path.append(os.path.dirname(os.path.abspath(__file__)))
-         14 #importing odd melt/fluid functions
-
-
-    ImportError: cannot import name 'Isotropy' from 'santex' (/home/sinan/.local/lib/python3.10/site-packages/santex/__init__.py)
-
-
-Let's try to make a contour figure to plot the effect of pressure and temperature on different compositional environment, namely Lherzolite, Harzburgite and Lherzolite with 8% Phlogopite.
-
-
-```python
-#Setting the temperature and pressure arrays
-temp = np.arange(600,1500,10) #setting up temperature array
-pressure = np.arange(1,6,0.1)
-#Creating a meshgrid
-T,P = np.meshgrid(temp,pressure)
-#flattening the arrays to load into pide
-T_array = T.ravel()
-P_array = P.ravel()
-
-p_obj = pide.pide() #creating the initial object
-
-#Setting the temperature and pressure arrays.
[…]
