+++ docs/source/7_Metasomatism_Maps_from_MT.md
@@ -1,331 +1,0 @@
-## <span style="color:green"> Notebook - Mapping metasomatised domains from MT Models</span>
-<span style="color:purple">Sinan Ozaydin, School of Geosciences, The University of Sydney, NSW 2006, Australia <br/> </span>
-<span style="color:purple">sinan.ozaydin@protonmail.com | sinan.ozaydin@sydney.edu.au </span>
-
-In this notebook, we will explore the calculation of first-order metasomatic addditions from inputted MT models. For this exercise, we are going to use the model of [Bedrosian et al., 2021](https://doi.org/10.1029/2021GL092970) and cropped temperature model (WINTERC-G) of [Fullea et al., 2021](https://doi.org/10.1093/gji/ggab094). 
-
-
-```python
-import os
-from pathlib import Path
-import numpy as np
-from scipy.interpolate import griddata
-import matplotlib.pyplot as plt
-from netCDF4 import Dataset
-
-import pide
-from pide.mt.mt_model_read import read_ModEM_rho
-from pide.utils.utils import read_csv, associate_coordinates
-from pide.utils.gis_tools import lat_lon_to_utm, get_utm_zone_number
-
-#Defining notebook path in the machine that is running.
-notebook_path = Path().resolve()
-```
-
-First of all, let's read the MT model by using the ***read_ModEM_rho*** function and determine the model center in UTM coordinates to georeference the model using the functions from ***pide.utils.gis_tools***. We will be using the depths according to 49 and 96 kilometer slices to calculate our metasomatism maps.
-
-
-```python
-wus_rho_path = os.path.join(notebook_path,'..','example_data','WUS_MT_Model','WUS.MT.Bedrosian2021.resistivity.rho')
-
-#model centers for georeferencing
-mc_lat = 40.75
-mc_lon = -113.25
-
-utm_no = get_utm_zone_number(mc_lon)
-mc_x, mc_y = lat_lon_to_utm(mc_lat,mc_lon)
-
-#reading the MT model file
-rho, mesh_x, mesh_y, mesh_z = read_ModEM_rho(rho_file_path=wus_rho_path)
-
-#converting x and y to utm coordinates
-x = mesh_x
-y = mesh_y
-
-#Getting the vertical indexes of 49 and 96 kilometer slices.
-idx_49 = (np.abs((mesh_z/1e3) - 49.0)).argmin()
-idx_96 = (np.abs((mesh_z/1e3) - 96.0)).argmin()
-
-#Getting the resistivities at 50 and 100 km
-rho_49 = rho[idx_49].flatten()
-rho_96 = rho[idx_96].flatten()
-
-cond_49 = 1.0/rho_49
-cond_96 = 1.0/rho_96
-
-```
-
-    /home/sinan/src/SEL/pide/utils/gis_tools.py:28: FutureWarning: This function is deprecated. See: https://pyproj4.github.io/pyproj/stable/gotchas.html#upgrading-to-pyproj-2-from-pyproj-1
-      utm_x, utm_y = transform(wgs84, utm, longitude, latitude)
-
-
-Now, let's read the temperature model with the help of utils.read_csv function. and then georefence them into the UTM coordinates.
-
-
-```python
-#Reading the temperature model
[…]
