+++ docs/source/Combined_Notebooks.md
@@ -1,0 +1,3437 @@
+![Example Image](../../docs/figures/pide_logo.png)
+## <span style="color:green"> Notebook - Starting with pide </span>
+<span style="color:purple">Sinan Ozaydin, School of Geosciences, The University of Sydney, NSW 2006, Australia <br/> </span>
+<span style="color:purple">sinan.ozaydin@protonmail.com</span>
+
+pide is a python3 library to calculate geophysical observableslike electrical conductivity and seismic velocity for the given compositional and thermodynamic environment. To derive this information it uses library of experimental models on electrical conductivity and thermoelastic constants. pide is constituted by three main classes:
+
+<span style="color:red;">-pide:</span> This is the class where calculations related to electrical conductivity and seismic velocities are performed. pide object is central to the library and all material and model classes utilizes pide object to calculate electrical conductivities and seismic velocities. 
+
+<span style="color:red;">-material:</span> A material holder class where the user can define a compositional environment and choices relevant to how the geophysical observables are going to be calculated. This can be done in the pide class by directly setting these parameters on a pide class, however, material class should create a tidy and interoperable environment for more complex calculations that can be streamlined through other processes.
+
+<span style="color:red;">-model:</span> The model class represents a collection of materials that are geospatially indexed and thermodynamically defined. The users can create model objects from informations derived from thermomechanical numerical (like underworld or aspect outputs), or geological models.
+
+**Other tools included in the toolkit of pide are:**
+
+<span style="color:green;">-geodyn:</span> A collection of scripts and methods to deal with input/output/processing of thermomechanical models.
+
+<span style="color:green;">-gravity and mag:</span> A collection of methods to calculate gravity and magnetic anomalies from given 3D environment set by a model object.
+
+<span style="color:green;">-mt:</span> A collection of tools to convert the outputs of pide into MT input scripts for forward and inverse modelling of these synthetic models. The codes currently implemented are ModEM and MARE2DEM.
+
+<span style="color:green;">-rheology:</span> A collection of methods to perform rheological calculations from the outputs of pide (e.g., mantle viscosity from MT driven water contents).
+
+<span style="color:green;">-inversion:</span> A collection of methods to invert for compositional parameterisations used in pide (e.g., mineral/melt contents, interconnectivity of phases, water contents) from real data.
+
+To understand the functions used in pide, one can utilise the following workflow:
+
+
+```python
+import pide
+
+#creating the pide object
+p_object = pide.pide()
+
+#Now, we can list all the methods used in pide class by:
+p_object.list_methods()
+
+
+```
+
+    - calculate_arrhenian_single
+    - calculate_bulk_mantle_water_solubility
+    - calculate_conductivity
+    - calculate_density_solid
[…]

[…]
