### 2024-10-07 sinanozaydin/pide#9 @sinanozaydin
readthedocs for pide are generated and online at the address: https://pide.readthedocs.io/
