### 2024-10-07 cd1e690
updates on doctstrings
