+++ pide/pide.py
@@ -64,7 +64,6 @@ warnings.filterwarnings("ignore", category=RuntimeWarning) #ignoring many Runtim
    \ \_\                     
 	\/_/                     
 """
-#version 0.2, May, 2024
 #pide - (P)etrophysical (I)nterpretation tools for geo(D)ynamic (E)xploration
 #initially developed by Sinan Ozaydin (University of Sydney, School of Geosciences
 #sciences, Sydney, Australia).
@@ -546,13 +545,17 @@ class pide(object):
 		value of 1.
 		
 		Input:
-		dict/float: mineral_name 
+		dict/float or 1-D array: mineral_name 
 		
 		Mineral_names:
-		ol,opx,cpx,garnet,mica,amp,quartz,plag,kfelds,sulphide,graphite,sp,rwd_wds,perov,mixture,other.
+		ol (olivine),opx (orthopyroxne),cpx (clinopyroxene),garnet,mica,amp (amphibole),quartz,
+		plag (plagioclase),kfelds (k-feldspar),sulphide,graphite,sp (spinel),
+		rwd_wds (ringwoodite or wadsleyite),perov (perovskite),mixture (mineral mixtures),
+		other (other minerals).
 		
 		Example:
 		set_composition_solid_mineral(ol = 0.6, opx = 0.4)
+		set_composition_solid_mineral(ol = [0.6,0.4], opx = [0.4,0.6])
 		
 		Organizes:
 		mineral fractions.
@@ -638,13 +641,14 @@ class pide(object):
 		value of 1.
 		
 		Input:
-		dict/float: rock_name 
+		dict/float or 1D-array: rock_name 
 		
 		Rock names:
 		granite,granulite,sandstone,gneiss,amphibolite,basalt,mud,gabbro,other_rock.
 		
 		Example:
 		set_composition_solid_rock(granite = 0.6, granulite = 0.4)
+		set_composition_solid_rock(granite = [0.6,0.4], granulite = [0.4,0.6])
 		
 		Organizes:
 		rock fractions.
@@ -700,7 +704,7 @@ class pide(object):
 		"""A method to set temperature of the environment.
 		
 		Input:
-		float: T || in Kelvin
+		float or 1-D array: T || in Kelvin
 		
 		Example:
 		set_temperature(temperature_array_list)
@@ -729,7 +733,7 @@ class pide(object):
 		"""A method to set pressure of the environment.
 		
 		Input:
-		float: P || in GPa
+		float or 1-D Array: P || in GPa
 		
 		Example:
 		set_pressure(pressure_array)
@@ -770,7 +774,7 @@ class pide(object):
 		"""A method to set depth of the environment.
 		
 		Input:
-		float: depth || in km
+		float or 1-D array: depth || in km
 		
 		Example:
 		set_depth(depth_array)
@@ -794,7 +798,7 @@ class pide(object):
 		water solubility models of olivine
 
 		Input:
-		dict/int: mineral_name = int
+		dict/int: mineral_name
 
 		Available minerals and assoicated index values:
 		ol:  0-Withers2012, 1-Bell2003, 2-Paterson1980 or 3-Default'
@@ -1070,7 +1074,10 @@ class pide(object):
 		str: mineral_name
 
 		mineral_names:
-		ol,opx,cpx,garnet,mica,amp,quartz,plag,kfelds,sulphide,graphite,sp,rwd_wds,perov,mixture,other.
+		ol (olivine),opx (orthopyroxne),cpx (clinopyroxene),garnet,mica,amp (amphibole),quartz,
+		plag (plagioclase),kfelds (k-feldspar),sulphide,graphite,sp (spinel),
[…]
