### 2024-10-07 sinanozaydin/pide#7 @sinanozaydin
Thank you @kujaku11 for the detailed review.

Here are the replies to your comments:

# General Comments
- Docstrings: They are updated and should be on every functions now.
- Tests: Tests suite is implemented for the core functionalities of the pide.
- Documentation: Documentation webpage is generated and are accesible via: https://pide.readthedocs.io/
- Notebooks: We have not addressed the binder suggestion, but we will likely to implement this in the near future.

# Manuscript
Minor comments are all addressed in the new version of paper.md. Some of them that requires clarification:

> Could add a second figure from one of the example notebooks, like from notebook 10 showing a 2D subsurface model that would highlight expected outputs based on the workflow.

We have addressed this issue by adding Figure 2, which takes the examples from Notebook 10 as suggested.

# Code Base

## Installation
We recognized that this problem was arising from the dependency SanTex when it is trying to install orix dependency. This also only happens if you are trying to install pide with pip on Windows (macOS and linux does not generate the same problem for some reason). New version of SanTex does not automatically installs orix for pip, but the users of windows are suggested to install the orix package through a conda package manager. So, the "pip install pide" command should now work without any hassle.

## Documentation

- We have generated a readthedocs page that includes the general information and the markdown version of the notebooks.
- Methods and docstring are described in more detail. We will be aware of any lack of it in the future as well.

## Tests

We have prepared a tests suite to address this issue.

# Examples

All notebooks should be working now and we addressed all the comments other than the potential field notebook.
We will be working on implementing potential fields on the next version if the reviewer can agree with this.

@kthyng
