### 2024-10-08 sinanozaydin/pide#3 @sinanozaydin
We have been able to recreate this issue and resolved it by fixing the dependency of orix as optional in pip (requires pycifrw) in SanTex library (a library we also author). This issue only seems to be happening at Windows computers.
