### 2024-10-23 845458f
better initial guess algorithm for inversion
now runs faster!
