+++ pide/inversion.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 
 import numpy as np
+import ipdb
 
 def _comp_adjust_(_comp_list, comp_alien, comp_old,final = False):
 
@@ -15,17 +16,24 @@ def _comp_adjust_(_comp_list, comp_alien, comp_old,final = False):
 	
 	return comp_list
 
-def _solv_cond_(index, cond_list, object, param, upperlimit, lowerlimit, search_increment, acceptence_threshold, 
-	init_guess = None, transition_zone = False, water_solv=False, comp_solv=False, comp_type = None, comp_index = None, low_value_threshold = None):
+def _solv_cond_(index, cond_list, object, param, upperlimit, lowerlimit, search_increment, acceptence_threshold, results_list = None,
+	init_guess = None, transition_zone = False, water_solv=False, comp_solv=False, comp_type = None, comp_index = None, low_value_threshold = None,):
 	
-	"""Simple grid-search solver to solve conductivities. This function should not be called directly.	
+	"""Simple line-search solver to solve conductivities. This function should not be called directly.	
 	"""
-		
-	param_search_array = np.arange(lowerlimit[index], upperlimit[index], search_increment)
 	
+	if results_list is not None:
+		if len(results_list) == 0:
+			init_guess = None
+		else:
+			init_guess = results_list[-1]
+			
+	param_search_array = np.arange(lowerlimit[index], upperlimit[index] , search_increment)
+		
 	if len(param_search_array) == 1:
 	
 		sol_param = upperlimit[index]
+		
 		if comp_solv == True:
 			if comp_type == 'mineral':
 				_comp_list = [object.quartz_frac[index], object.plag_frac[index], object.amp_frac[index], object.kfelds_frac[index], object.opx_frac[index], object.cpx_frac[index],
@@ -65,13 +73,23 @@ def _solv_cond_(index, cond_list, object, param, upperlimit, lowerlimit, search_
 
 		restart = True
 		init_search_increment = np.array(search_increment)
+		init_restart = True
 		
 		while restart:
 			
 			restart = False
 			
-			for j in range(0,len(param_search_array)):
+			if init_guess == None:
+				idx_start_search = 0
+			else:
+				if init_restart == True:
+					idx_start_search = np.argmin(np.abs(param_search_array-init_guess))
+					init_restart = False
+				else:
+					idx_start_search = 0
 			
+			for j in range(idx_start_search,len(param_search_array)):
+				
 				if comp_solv == True:
 				
 					if comp_type == 'mineral':
@@ -89,7 +107,7 @@ def _solv_cond_(index, cond_list, object, param, upperlimit, lowerlimit, search_
 						
 						if comp_type == 'mineral':
 							object.mineral_frac_list[idx_t][index] = comp_list[idx_t]
-						elif comp_type == 'mineral':
+						elif comp_type == 'rock':
 							object.rock_frac_list[idx_t][index] = comp_list[idx_t]
 						
 					exec('object.' + param + '[' + str(index) + ']='  + str(param_search_array[j]))
@@ -102,7 +120,7 @@ def _solv_cond_(index, cond_list, object, param, upperlimit, lowerlimit, search_
 				else:
 				
 					exec('object.' + param + '[' + str(index) + ']='  + str(param_search_array[j]))
-				
+					
 				if water_solv == True:
 					if transition_zone == False:
 						object.mantle_water_distribute(method = 'index', sol_idx = index)
@@ -119,29 +137,16 @@ def _solv_cond_(index, cond_list, object, param, upperlimit, lowerlimit, search_
 						sol_param = param_search_array[j]
 					else:
 						if param_search_array[j] < low_value_threshold:
-						
 							sol_param = 0.0
-							
[…]
