### 2024-10-30 b392c08
fixing stuff in rheology
