+++ pide/mt/mt_model_conversion.py
@@ -119,7 +119,7 @@ def convert_2DModel_2_MARE2DEM(file_out, conductivity_array, mesh, boundaries = 
 	print('--------- ' + str(boundaries['bottom']) + '---------')
 	
 	
-def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, core_bounds = None, station_array = None, **kwargs):
+def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, starting_index = None, first_layer_resistivity = None, core_bounds = None, station_array = None, **kwargs):
 
 	#This is a function that converts the constructed 3D geodynamic model into data points
 	#that can be readable by the 3D MT modelling algorithm ModEM.
@@ -140,6 +140,7 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, core_bounds = No
 	"""
 	
 	from scipy.interpolate import griddata
+	import ipdb
 	
 	# core_mesh_size = kwargs.pop('core_mesh_size', mesh[0][0][1][0] - mesh[0][0][0][0])
 	num_horiz_bounds = kwargs.pop('num_horiz_bounds', 8)
@@ -159,7 +160,7 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, core_bounds = No
 				rho[i][j] = conductivity_array[(i*slice_len)+j]
 			except IndexError:
 				raise IndexError('The mesh structure entered does not match the conductivity array. Be sure the entered format mesh = (x_mesh_centers,y_mesh_centers,z_mesh_centers) in tuples are correct.')
-					
+	
 	#determining the indexes
 	start_index_list = []
 	sea_index_list = []
@@ -170,10 +171,18 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, core_bounds = No
 			sea_index_list.append(i)
 	
 	sea_index = sea_index_list[-1]
+
+	#getting rid of the all air layers.
+	if starting_index is None:
+		starting_index = start_index_list[0]
+	else:
+		starting_index = starting_index
+		
+	z_mesh = z_mesh[:starting_index]
 	
 	for i in range(0,len(rho)):
 		if i > sea_index:
-			rho[i][(rho[i] == -999.0)] = 1e-14
+			rho[i][(rho[i] == -999.0)] = -999
 	
 	outs = np.float64(x_mesh[1] - x_mesh[0])
 	xy_out = []
@@ -181,7 +190,7 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, core_bounds = No
 		outs = outs * horiz_bound_incr
 		xy_out.append(outs)
 	xy_out = np.array(xy_out)
-		
+	
 	outs_y = np.float64(y_mesh[1] - y_mesh[0])
 	yx_out = []
 	for i in range(0, num_horiz_bounds):
@@ -194,7 +203,7 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, core_bounds = No
 	for i in range(0, num_vert_bounds):
 		outs_z = outs_z * vert_bound_incr
 		z_out.append(outs_z)
-	outs_z = np.array(outs_z)
+	z_out = np.array(z_out)
 	
 	x_core = np.ones(len(x_mesh)) * np.float64(x_mesh[1] - x_mesh[0])
 	y_core = np.ones(len(y_mesh)) * np.float64(y_mesh[1] - y_mesh[0])
@@ -223,10 +232,12 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, core_bounds = No
 	
 	rho_new = np.array(rho_new)
 	
+	
+	
 	for i in range(0,num_vert_bounds):
[…]
