### 2024-11-10 5fd0836
a little coding for reorganisation script
