### 2024-11-11 93c6681
some edits to create modem files
