+++ pide/mt/mt_model_conversion.py
@@ -486,10 +486,16 @@ def get_station_elevation_ModEM_rho(input_rho, station_location_arrays = None, a
 			
 	return (station_location_arrays[0]*1e-3,station_location_arrays[1]*1e-3,np.array(depth_sol)*1e-3)
 		
-def reArrangeModEMdatfile(file_out, input_dat, err_function ,station_skip = None, station_selection_list = None):
+def reArrangeModEMdatfile(file_out, input_dat, err_function ,err_array, station_skip = None, station_selection_list = None):
 
 	from pide.utils.utils import read_csv
-
+	
+	err_methods = ["off-diagonal","diagonal","independent","X-based"]
+	
+	if (err_function in err_methods) == False:
+	
+		raise KeyError(f'The error function has to be one of the following string values: {str(err_function)}')
+		
 	dat_data = read_csv(filename = input_dat, delim = ' ')
 
 	dash_found = False
@@ -510,18 +516,191 @@ def reArrangeModEMdatfile(file_out, input_dat, err_function ,station_skip = None
 
 	station_names = []
 	station_change_row = []
+	station_x = []
+	station_y = []
 	for row in range(start_idx,limitlines):
 		if dat_data[row][1] != dat_data[row-1][1]:
 			station_names.append(dat_data[row][1])
+			station_x.append(float(dat_data[row][4]))
+			station_y.append(float(dat_data[row][5]))
 			change = row
 			station_change_row.append(change)
 			lenst = len(station_change_row)
 	station_change_row.append(limitlines)
 
-	import ipdb
-	ipdb.set_trace()
-
+	lines = []
+	for i in range(0,station_change_row[0]):
+		lines.append(' '.join(dat_data[i]) + '\n')
+		
 	if station_skip is not None:
+	
 		for i in range(0,len(station_change_row),station_skip):
+			
+			try:
+				for j in range(station_change_row[i],station_change_row[i+1]):
+					
+					if dat_data[j][7] == "ZXX":
+						
+						if err_function == "off-diagonal":
+							err = '% .5E' %  (np.sqrt((float(dat_data[j+1][8]) + float(dat_data[j+1][9])*1j) * (float(dat_data[j+2][8]) + float(dat_data[j+2][9])*1j)) * err_array[0] * 1e-2)
+							lines.append('  '.join(dat_data[j][:10]) + "  " + err + '\n')
+						elif err_function == "diagonal":
+							err = '% .5E' %  (np.sqrt((float(dat_data[j][8]) + float(dat_data[j][9])*1j) * (float(dat_data[j+3][8]) + float(dat_data[j+3][9])*1j)) * err_array[0] * 1e-2)
+							lines.append('  '.join(dat_data[j][:10]) + "  " + err + '\n')
+						elif err_function == "independent":
+							err = '% .5E' %  ((float(dat_data[j][8]) + float(dat_data[j][9])*1j) * err_array[0] * 1e-2)
+							lines.append('  '.join(dat_data[j][:10]) + "  " + err + '\n')
+						elif err_function == "X-based":
+							err = '% .5E' %  (np.sqrt((float(dat_data[j][8]) + float(dat_data[j][9])*1j) * (float(dat_data[j+1][8]) + float(dat_data[j+1][9])*1j)) * err_array[0] * 1e-2)
+							lines.append('  '.join(dat_data[j][:10]) + "  " + err + '\n')
+							
+					elif dat_data[j][7] == "ZXY":
+						
+						if err_function == "off-diagonal":
[…]
