### 2024-11-13 6fd0a0d
Minor fix to Harmonica BibTeX reference
Put `Fatiando a Terra Project` inside curly braces so BibTeX doesn't
render it as a person name.
