+++ docs/paper/paper.bib
@@ -53,7 +53,7 @@
   publisher={Oxford University Press}
 }
 @software{fatiando_a_terra_project_2024,
-  author       = {Fatiando a Terra Project and
+  author       = {{Fatiando a Terra Project} and
                   Castro, Yago M. and
                   Esteban, Federico D. and
                   Li, Lu and
