### 2024-11-13 sinanozaydin/pide#13 @santisoler
Thanks @sinanozaydin for addressing all my previous comments. ~I'll close this issue now.~

I'll also leave some minor comments on the latest version of the paper (93c6681e975e1ad00412c06cbb11da7f76076d2a).

## Figure 1

In Figure 1 the `inversion.py` module still has a trailing `.py`. I think it would be nice to remove it to be consistent with how the other modules are referred to in the figure.

Moreover, the blocks listed in Outputs are usually arrays of some sort? If so, I'd recommend changing the legend to _arrays_ or _vectors_.

## Paragraph

The last two sentences of the third paragraph in *Library modules and methods* reads:

> This class allows the user to define some parameters tied to other parameters via user-defined functions. This allows the users to streamline some specific processes in a modular manner.

I think there's significant redundancy in those. I'd suggest the authors to revisit them.

## Minor fix to Harmonica reference

I opened #17 to fix how the citation to Harmonica is done. By putting `Fatiando a Terra Project` inside curly braces, BibTeX recognizes as an _institutional_ author and not a real person. Feel free to comment on #17 about this.

## Extend text regarding inversions

The last paragraph of the article mentions some of the capabilities of the `inversions` module:

> inversion modules, on the other hand, can be utilised to invert for specific input parameters
(e.g., composition, melt content, mineral interconnection) that fit real geophysical data.

I think the description of such module is quite vague. I'd recommend the authors to extend the description of the module, since to _"invert for parameters that fit real geophysical data"_ could comprehend a wide variety of inversion methods (voxel inversions, parametric inversion, single-parameter optimization, etc).

---

~These are minor comments that won't block the acceptance for publication. Nonetheless, I'd recommend the authors to address them in order to improve the quality of the submission.~

> [!NOTE]
> **Updated 2024-11-13**: striked out comments on closing this issue, and this suggestions not being required for publication.
