### 2024-11-13 sinanozaydin/pide#11 @santisoler
Thanks @sinanozaydin for adding the contributing guidelines. They are greatly appreciated.
