### 2024-11-13 b324c1b
fixes to santisolers
remove ipdb calls
