+++ pide/mt/mt_model_conversion.py
@@ -32,8 +32,6 @@ def convert_2DModel_2_MARE2DEM(file_out, conductivity_array, mesh, boundaries = 
 	else:
 		raise ValueError('Please enter a valid response for cond_unit: "conductivity" or "resistivity".')
 		
-	
-		
 	boundary_list = ['left','right','top','bottom']
 	#Defaulting into mesh boundaries if w
 	if boundaries == None:
@@ -119,7 +117,7 @@ def convert_2DModel_2_MARE2DEM(file_out, conductivity_array, mesh, boundaries = 
 	print('--------- ' + str(boundaries['bottom']) + '---------')
 	
 	
-def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, starting_index = None, first_layer_resistivity = None, core_bounds = None, station_array = None, **kwargs):
+def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, scramble_first_layer = True, starting_index = None, first_layer_resistivity = None, station_array = None, **kwargs):
 
 	#This is a function that converts the constructed 3D geodynamic model into data points
 	#that can be readable by the 3D MT modelling algorithm ModEM.
@@ -129,8 +127,10 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, starting_index =
 	file_out: filename or full directory to output the result. Includes the format (e.g., csv)
 	conductivity_array: calculated conductivity field in np.ndarray()
 	mesh: mesh of the associated conductivity field.
-	core_bounds: boundaries dictionary in  {'left','right','top','bottom'}	
-	
+	scramble_first_layer: boolean input to part the first layer automatically into multiple layer accomodate for MT modelling
+	starting_index: overriding index value to set the layer denoting to 0. Setting this will scrape all the structure below this index.
+	first_layer_resistivity: Resistivity value in ohm m to overwrite the resistivity of the first layer, mimicking the sedimentary layer with water table.
+		
 	Keyword Arguments:
 	core_mesh_size:
 	num_horiz_bounds:
@@ -140,7 +140,6 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, starting_index =
 	"""
 	
 	from scipy.interpolate import griddata
-	import ipdb
 	
 	# core_mesh_size = kwargs.pop('core_mesh_size', mesh[0][0][1][0] - mesh[0][0][0][0])
 	num_horiz_bounds = kwargs.pop('num_horiz_bounds', 8)
@@ -212,7 +211,7 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, starting_index =
 	x_out = np.concatenate([xy_out[::-1],x_core,xy_out])
 	y_out = np.concatenate([yx_out[::-1],y_core,yx_out])
 	z_out = np.concatenate([z_core,z_out])
-			
+	
 	rho_new = []
 	
 	for i in range(0,len(z_mesh)):
@@ -231,8 +230,6 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, starting_index =
 		rho_new.append(np.array(rho_local))
 	
 	rho_new = np.array(rho_new)
-	
-	
 	
 	for i in range(0,num_vert_bounds):
 		rho_new = np.insert(rho_new,0,np.ones(len(rho_new[-1])) * np.nan, axis = 0)
@@ -302,6 +299,39 @@ def convert_3DModel_2_ModEM(file_out, conductivity_array, mesh, starting_index =
 
 	if first_layer_resistivity is not None:
[…]
