### 2024-11-13 sinanozaydin/pide#14 @santisoler
I agree with your decision, I think it's the most sensible one considering the current stage of the project.

I'd recommend to leave this Issue open in case the developers desire to apply these changes in the future.
