### 2024-11-13 sinanozaydin/pide#9 @santisoler
Thanks @sinanozaydin for addressing my previous comments. I think having the documentation of `pide` in RTD is a great step forward. I'll therefore close this issue.

Considering the amount of functionalities that live inside `pide` (it currently has 7 submodules containing many functions and classes), I **heavily encourage** to create an API reference listing all functions, classes and their modules.
I don't think the lack of an API reference should block the acceptance of `pide` for publication in JOSS, therefore I'm not going to require it for publication.
