### 2024-11-13 sinanozaydin/pide#10 @santisoler
Thanks for adding a test suite. Looking forward to see the instructions on the documentation pages.

---

I would suggest to keep extending the test suite to make sure more pieces of the code are being tested.

Another recommendation I could give is to add instructions to get a coverage report that can be used to identify which lines of code are not being tested. This can be done using [`coverage`](https://coverage.readthedocs.io) or the [`pytest-cov`](https://github.com/pytest-dev/pytest-cov) plugin.

> [!NOTE]
> These are recommendations and it won't block the publication of `pide` in JOSS.

---

I'm closing this issue.
