### 2024-11-13 sinanozaydin/pide#20 @santisoler
Add instructions to seek support
> [!NOTE]
> This issue was opened as part of the JOSS review process: https://github.com/openjournals/joss-reviews/issues/7021

## The issue

JOSS requires that contributing guidelines includes instructions on how to seek support when using the submitted software:

> **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> _Source_: [JOSS reviewer checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#documentation)

Currently `pide` has a [`CONTRIBUTING.md`](https://github.com/sinanozaydin/pide/blob/main/CONTRIBUTING.md) file that includes instructions on how to contribute to the repository and how to report issues or problems, but it lacks instructions on how to seek support while using it.

## Possible solution

Include instructions on how to get in touch with the developers in the `CONTRIBUTING.md`.
