### 2024-11-13 sinanozaydin/pide#13 @santisoler
After reviewing this issue, I think there are a few points in my previous comment that worth pushing forward for the JOSS publication. Particularly related to the _Quality of writing_ item in the [reviewer checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper).

So I'll reopen this issue to keep track of it.
