### 2024-11-13 #7021 @santisoler
# Summary of second pass of my review

I would like to thank @sinanozaydin for all the effort put into addressing all my previous comments. I would also like to apologize again for my delays on providing replies to the requested changes.

## Major issues

After taking a second pass to the submitted article and repository I found a few issues that should be addressed before it can be accepted for publication.

### Improvements to the paper

During the first pass to the review I opened https://github.com/sinanozaydin/pide/issues/13 to track improvements to the paper. During this second pass I mentioned a few places for improvement. In summary:

- [ ] Minor fix to Figure 1
- [ ] Improvements to final sentences of paragraph.
- [ ] Minor fix on one of the references (I opened https://github.com/sinanozaydin/pide/pull/17 with a suggested fix).
- [ ] Extend description of the inversion capabilities of `pide`.

### Contributing guidelines

@sinanozaydin successfully added a [`CONTRIBUTING.md`](https://github.com/sinanozaydin/pide/blob/main/CONTRIBUTING.md) file that includes guidelines for contributing to the repository and to report issues. Nonetheless, it lacks instructions on how to seek support, a requirement listed in the JOSS reviewer checklist. More information can be found in:

- [ ] https://github.com/sinanozaydin/pide/issues/20

### Typos in notebooks

During my first review I noticed a few typos in the notebooks provided in the `pide` repository. I opened a PR to highlight them and fix them. I recommend the authors to revisit them:

- [ ] Merge https://github.com/sinanozaydin/pide/pull/8 (or fix typos).

### Fix missing comma in `setup.py`

After adding `psutil` to the list of dependencies in `setup.py` (https://github.com/sinanozaydin/pide/pull/16), a missing comma in the previous dependency generated an error while trying to install `pide` with `pip`. I opened a PR to solve this issue:

- [ ] Merge https://github.com/sinanozaydin/pide/pull/19 (or apply similar changes to address this issue).

### Add instructions on how to run the tests

After opening https://github.com/sinanozaydin/pide/issues/10, the authors added tests to `pide` that can be automatically run. As mentioned in the same issue, it would be nice to:

- [ ] Add instructions on how to run the tests to the documentation pages.

## Additional recommendations

I would like to acknowledge all the hard work that the authors put in developing `pide`. By taking a look at the amount of code and features that `pide` includes, we can easy tell all the effort that is required to put it together.

I also think that `pide` has a few places of improvements that could help to increase the quality of the software. The following suggestions are not required for accepting `pide` for publication, although I strongly suggest the authors to consider them for the long-term goals of `pide`.

> [!IMPORTANT]
> The following suggestions are not required to accept `pide` for publication.

### API Reference

Considering the amount of functions, classes and methods that live inside `pide`, I think users will greatly appreciate having a documented API reference. This could be done with Sphinx and some of its plugins, like `autodoc` and `autosummary`, or `autodoc2` and `myst_parser` (since `pide` is using the Myst Sphinx extension to build the docs).

Given the current stage of `pide`, I don't consider it a strong requirement to accept `pide` for publication, but I **strongly suggest** the authors to add an API reference in the near future.

See comment in https://github.com/sinanozaydin/pide/issues/9 for reference.

### Use a standard style for the docstrings

Documenting functions and methods is always welcomed in any form. Although, making use of a standard style for writing the documentation greatly improves how the docstrings are read and also parsed by documentation building plugins (like Sphinx's `autosummary`).

For example, the [`numpydoc` style](https://numpydoc.readthedocs.io/en/latest/format.html) is one of the most used among scientific libraries.

This is related to the previous point: using a documentation style makes it easier to build an API reference in the documentation pages.

### Code style and quality

I think there are places to improve the code quality of `pide`.

For example, the usage of `eval` is usually discouraged unless it's strictly necessary. The following lines

# https://github.com/sinanozaydin/pide/blob/b324c1b0eb9219f5519e7201b4ee6a5f1154c42a/pide/inversion.py#L310-L313
if idx_set == comp_index:
  eval(f"object.set_parameter('{param_name}',c_list)")
else:
  eval(f"object.set_parameter('{min_list[idx_set]}',comp_list[idx_set])")
```

could be replaced by calls to the `set_parameter` method:

if idx_set == comp_index:
    object.set_parameter(param_name, c_list)
else:
    object.set_parameter(min_list[idx_set], comp_list[idx_set])
```

Currently, the Python ecosystem offers some tools to automatically detect some common mistakes (like unused imports, arguments, and variables, comparisons against bools in `if` statements, etc), like [`ruff`](https://github.com/astral-sh/ruff) and [`flake8`](https://flake8.pycqa.org).

Moreover, styling our Python code in a standard way improves readability. Manually formatting the code style could be tedious, that's why auto-formatters like [Black](https://black.readthedocs.io/en/stable/index.html) are widely used in Python projects.

I would suggest the authors to start using these tools to standardize the code style of `pide` and to improve its code quality.

### Extend test suite

After opening https://github.com/sinanozaydin/pide/issues/10, the authors added tests to `pide` that can be automatically run. I consider the implemented test suite to be sufficient to accept `pide` for publication. Nonetheless, I suggest the authors to work on extending it to ensure that most of the code in `pide` is being tested, and to ensure that the code works as expected even on corner cases.

### PEP compliance

I opened two issues (https://github.com/sinanozaydin/pide/issues/14 and https://github.com/sinanozaydin/pide/issues/18) suggesting the authors to make `pide` to comply with PEP8 and PEP518. The former states how to choose variable names, while the latter introduces the usage of `pyproject.toml` in favor of `setup.py` and `setup.cfg`.

I would recommend the authors to revisit these issues in the future.
