+++ CONTRIBUTING.md
@@ -7,6 +7,10 @@
 
 If you find a bug, please file an issue on the repository with detailed information on how to reproduce the issue.
 
+### Seeking Support
+
+If you are having any problems while trying to use *pide*, please open an issue on the GitHub page.
+
 ### Proposing Features
 
 If you'd like to propose a new feature, please open an issue and start a discussion. Once the community and maintainers approve, you can submit a pull request.
