### 2024-11-14 sinanozaydin/pide#20 @sinanozaydin
Thanks for this suggestion! We introduced a new section to the `CONTRIBUTING.md` file.
