### 2024-11-14 32086d2
notebook typo fixes
