+++ docs/source/Combined_Notebooks.md
@@ -3,7 +3,7 @@
 <span style="color:purple">Sinan Ozaydin, School of Geosciences, The University of Sydney, NSW 2006, Australia <br/> </span>
 <span style="color:purple">sinan.ozaydin@protonmail.com</span>
 
-pide is a python3 library to calculate geophysical observableslike electrical conductivity and seismic velocity for the given compositional and thermodynamic environment. To derive this information it uses library of experimental models on electrical conductivity and thermoelastic constants. pide is constituted by three main classes:
+pide is a python3 library to calculate geophysical observables like electrical conductivity and seismic velocity for the given compositional and thermodynamic environment. To derive this information it uses library of experimental models on electrical conductivity and thermoelastic constants. pide is constituted by three main classes:
 
 <span style="color:red;">-pide:</span> This is the class where calculations related to electrical conductivity and seismic velocities are performed. pide object is central to the library and all material and model classes utilizes pide object to calculate electrical conductivities and seismic velocities. 
 
@@ -201,6 +201,541 @@ The pide object can be reset to default values can be done as follows:
 
 ```python
 p_object.reset()
+```
+## <span style="color:green"> Notebook - Conversion of 2D Underworld Model II - Narrow Rift Inversion</span>
+<span style="color:purple">Sinan Ozaydin, School of Geosciences, The University of Sydney, NSW 2006, Australia <br/> </span>
+<span style="color:purple">sinan.ozaydin@protonmail.com | sinan.ozaydin@sydney.edu.au </span>
+
+In this notebook, we will convert a numerical model output from underworld to geophysical parameters, electrical conductivity and seismic velocity. First, let's import the functions we are going to use.
+
+
+```python
+import os
+from pathlib import Path
+
+import pide
+from pide.material import Material
+from pide.model import Model
+from pide.geodyn.read_uw_model import *
+from pide.geodyn.interpolate_fields import interpolate_2d_fields
+from pide.geodyn.plot_models import *
+from pide.geodyn.material_process import *
+from pide.geodyn.write_uw_model import write_2d_field_h5
+
+#setting up source folder of the files
+notebook_path = Path().resolve()
+source_folder = os.path.join(notebook_path,'..','example_data','uwconversion', 'NRift')
+```
+
+    /home/sinan/.local/lib/python3.10/site-packages/matplotlib/projections/__init__.py:63: UserWarning: Unable to import Axes3D. This may be due to multiple versions of Matplotlib being installed (e.g. as a system package and as a pip package). As a result, the 3D projection is not available.
+      warnings.warn("Unable to import Axes3D. This may be due to multiple versions of "
+
+
+Reading the data files of thermomechanical model outputs in h5 formats. They can be read by the ***read_h5_file*** function.
+
+
+```python
+#setting up filename folders for the h5 files.
+temp_fnm = os.path.join(source_folder,'temperature-152.h5')
+pstrain_fnm = os.path.join(source_folder,'projPlasticStrain-152.h5')
[…]
