### 2024-11-14 #7021 @sinanozaydin
Thanks a lot for the detailed reviews @santisoler! I am not particularly well-versed in the standards set by the software communities, so this was a very educational review process, and I thoroughly enjoyed it! Here are my replies to your comments:

# Improvements to the paper:
- Figure 1 is fixed as suggested.
- The final sentences of the third paragraph were removed.
- Fixes to references were implemented.
- We added extra information to the text regarding the inversion module and some future projects.

# Contributing Guidelines
- A guideline for running tests is implemented in the CONTRIBUTING.md file.

# Typos in notebooks
- Thanks a lot for the very detailed look on the notebooks! We have fixed these typos.

# Fix missing comma in setup.py
- We have fixed this issue as suggested.

# Additional recommendations
- API Reference: I have tried to put API references in the readthedocs but failed at my first try. I realised this was not an easy task to do if one uses an index.md instead of index.rst. In the near future, I will try to change this.

- Use a standard style for docstrings: I have really tried to do this but failed at it for some reason. When I have time, I will follow the numpydoc reference style for docstring generation.

- Code style and quality: I know the `eval` is frowned upon, but I could not really find another workaround for putting any string variable (`param_name`) set by the user. As far as I understand from some quick research, we should avoid using `eval` because it could potentially pose a security risk to performing operations. I think in our case, it should not host a security risk since it cannot simply be replaced by an `os` or `open()` command in this given example. Thanks a lot for other suggestions on how to make the code more elegant. I will have a look at these tools when I can.

# Extending test suite
- I just set up an agenda for myself to complete two tests every week. Hopefully, it will be finished in a couple of months.

# PEP compliance
- I will have a look at this transformation in the near future.
