### 2024-11-14 sinanozaydin/pide#22 @santisoler
Add instructions on how to run the tests to the documentation pages
I open this issue to keep track of the addition of instructions on how to run the tests into the documentation pages.

See #10 for reference.

> [!NOTE]
> This Issue was opened as part of the JOSS review process: https://github.com/openjournals/joss-reviews/issues/7021
