### 2024-11-14 sinanozaydin/pide#13 @santisoler
I'm closing this issue after my comments have been addressed.
