### 2024-11-14 #7021 @santisoler
Hi @sinanozaydin. Thanks for the swift replies to my recent comments, and for quickly start working to address them.

I wanted to let you know that I spotted a few extra typos in the notebooks (that I missed in https://github.com/sinanozaydin/pide/pull/8) and opened a PR to fix them: https://github.com/sinanozaydin/pide/pull/21

I also opened an issue suggesting to add instructions on how to run the tests to the documentation pages: https://github.com/sinanozaydin/pide/pull/22

After addressing these two issues, I'll consider `pide` to be ready for publication.

---

> API Reference: I have tried to put API references in the readthedocs but failed at my first try. I realised this was not an easy task to do if one uses an index.md instead of index.rst. In the near future, I will try to change this.

I understand. I agree that it might get tricky to add API reference using `myst_parser` in Sphinx. Also, docstrings with wrong format could cause issues when `autodoc` or `autosummary` tries to parse them to generate the HTML files for them. That's why I mentioned that using some standard style like `numpydoc` is connected to this.

> Code style and quality: I know the eval is frowned upon, but I could not really find another workaround for putting any string variable (param_name) set by the user. As far as I understand from some quick research, we should avoid using eval because it could potentially pose a security risk to performing operations. I think in our case, it should not host a security risk since it cannot simply be replaced by an os or open() command in this given example. Thanks a lot for other suggestions on how to make the code more elegant. I will have a look at these tools when I can.

I'm not sure if I fully understand the problem of replacing `eval` in that case. If `param_name` is a string variable, and you want to pass that string to the method, wouldn't calling the method with `param_name` as argument just work? Something like this:

```python
object.set_parameter(param_name, c_list)
```

You're right that using `eval` could lead to security issues, and that's one of the reasons it's discouraged to use it. In the case you mention, you can totally inject code through the `param_name` variable that could include calls to `os` or any other builtin library, as long as the string passed to `eval` in syntactically correct.

But the usage of `eval` is also discouraged for practical reasons as well: the code within the string cannot be linted by tools like `flake8`, `ruff`, or type checkers like `mypy`. Moreover, calls to `eval` could also impact performance since they have some overhead. Moreover, with the flexibility that Python offers, in most scenarios a call to `eval` can be replaced with actual Python code.

Even if `pide` doesn't need to be critically secure and the performance hits might be small, I think following best practices is the best path to follow when developing code that will be used by other people, not only it can be very educational.

---

Thanks for your comments. I also enjoyed reviewing `pide`!
