### 2024-11-15 #7021 @sinanozaydin
Thanks a lot @santisoler for your quick response! To answer the remaining comments:
- I merged your pull-request regarding the remaining typos.
- I have included the "Running Tests" section in the documentation page. It is quite simple right now, but I will extend and try to follow your suggestions in the next months while building new tests.
- Problem with `eval` in that code snippet is the fact that I want to give the user ability to change the parameter name, so they can invert for any paramter that can be defined in pide. Someone can define the param_name to be 'bulk_water' or 'mica_frac'. If `param_name` is entered like the way you mentioned, it wont work as intended.

Let me know anything else is needed. Thank you!
