### 2024-11-15 sinanozaydin/pide#22 @santisoler
@sinanozaydin started adding instructions on how to run the tests in https://github.com/sinanozaydin/pide/commit/75a759ded973a7284da0d8166b6c24281dfc61a9.

I think these are enough to meet JOSS requirements for publication. I'll leave this issue open in case the authors want to keep working on it.

Feel free to close it if you desire.
