### 2024-11-15 #7021 @santisoler
Hi @sinanozaydin. Thanks for addressing my comments.

@kthyng, I consider `pide` to be ready for publication.

---

## Comment on usage of `eval`

> [!NOTE]
> Just keeping the conversation around this topic. There's no action I expect to be done regarding the usage of `eval` as part of the review process.

> Problem with eval in that code snippet is the fact that I want to give the user ability to change the parameter name, so they can invert for any paramter that can be defined in pide. Someone can define the param_name to be 'bulk_water' or 'mica_frac'. If param_name is entered like the way you mentioned, it wont work as intended.

I think I disagree with the last sentence. In case the user sets `param_name="bulk_water"`, then the string that gets passed to the `eval` function will be:

```python
param_name = "bulk_water"
command = f"object.set_parameter('{param_name}',c_list)"
print(command)
```
```
object.set_parameter('bulk_water',c_list)
```

Passing this `command` as argument to `eval` as follows:

```python
eval(command)
```

produces the same result as calling the method as follows:

```python
object.set_parameter(param_name,c_list)
```

So, replacing the `eval` call by an actual call to the method would still allow users to any parameter provided by `param_name`.

Let me know what do you think. I'll be glad to elaborate on this if I'm not being clear enough.
