### 2024-11-15 sinanozaydin/pide#3 @kujaku11
@sinanozaydin This issue still persists.  Tried `pip install pide` on a windows machine.  Suggest adding a note in the install instructions for now if you are using windows need to be award of this issue. Or add `pycifrw` a dependency so it installs before `pide`
