### 2024-11-15 7e0bfa8
change santex version requirement
