### 2024-11-15 sinanozaydin/pide#3 @sinanozaydin
@kujaku11 I have updated setup.py to install specific version of santex, so it will force you to update. If this issue persists I will add a specific install requirements for windows.
