### 2024-11-15 sinanozaydin/pide#3 @sinanozaydin
It was working a month ago, but now I'm also having the same problem. I'll have to talk to developer of santex about this.
