### 2024-11-16 sinanozaydin/pide#7 @kujaku11
@kthyng The author has done a commendable job on updating the repository and fixing issues.  The author has added a test suite (@sinanozaydin for development, suggest adding a GitAction to automatically test on push), documentation through ReadTheDocs, and updated the example notebooks.  There are only 2 remaining issues #2 and #3.  Once the author has closed those issues, this should be ready for publication.
