### 2024-11-16 37ae002
fixed installation for windows machines
