### 2024-11-16 sinanozaydin/pide#3 @sinanozaydin
I put new installation instructions for windows machines in readme and readthedocs.
