### 2024-11-16 sinanozaydin/pide#7 @sinanozaydin
Thanks a lot for the detailed reviews @kujaku11 ! For the remaining issues you mentioned.

#2 - This issue should not happen if the main branch is utilised in installing `pide.`

#3 - I have added instructions on `README.md` and readthedocs how to install `pide` on a Windows computer.

We will be planning to include a more sophisticated approach on testing in the near future.
