### 2024-11-16 cbdcacc
changes to add badges
