### 2024-11-16 abfc6c0
some little updates to readme and index
