### 2024-11-23 #7021 @sinanozaydin
@kthyng I have addressed the last comments of @kujaku11 [here](https://github.com/sinanozaydin/pide/issues/7).
