### 2024-12-09 1e39a76
update on discarding commented out materials
