### 2024-12-10 e8b8b18
commented out melt_array thing
