### 2024-12-11 58d89ac
some additional changes to fluid conductivities that do not work below 373 K
