### 2024-12-11 6cfb1d2
fix to conductivity model selection exceptions
