### 2024-12-18 #7021 @kthyng
@kujaku11 Great to hear! Can you please revisit your review checklist and check off all boxes, assuming you consider them finished. Also if the issues you have raised are finish (https://github.com/sinanozaydin/pide/issues/2, https://github.com/sinanozaydin/pide/issues/3) it would be best for you to close them since you opened them. Alternatively if they are done enough for the review but still have outstanding future notes that aren't required for this review, feel free to leave them open but make a comment to that effect in each, as needed. Thank you!
