### 2024-12-18 sinanozaydin/pide#3 @kujaku11
This seems to be fixed, but I will keep this open as it doesn't seem to be stable yet. For now, it's good enough for the review process.
