### 2024-12-20 #7021 @kthyng
@sinanozaydin please copy the bulleted list about of Author tasks into a checklist and go through those items, reporting back as needed, and let me know when you're finished.
