+++ pide/__init__.py
@@ -1,5 +1,5 @@
 """
-pide v0.3
+pide v1.0
 
 """
 
@@ -8,6 +8,6 @@ import sys, csv, re, warnings, json
 from .pide import pide
 
 __author__ = 'Sinan Ozaydin'
-__version__ = '0.3'
+__version__ = '1.0.0'
 
 
