### 2025-01-08 bd9877e
fix to melt/fluid interconnectivity errors
