### 2025-01-08 #7021 @kthyng
paper comment @sinanozaydin: "earth" isn't consistently capitalized. Otherwise it looks good!
