### 2025-01-08 19e13b0
last fixes to the paper
all "earth" is capitalised
