### 2025-01-08 #7021 @sinanozaydin
Hello @kthyng,

I have fixed the capitalization of Earth on the github repo, should I push a new release on github/zenodo for this? This will create a new doi on Zenodo.
