### 2025-01-08 c3a3a49
last fixes to the paper
