### 2025-01-08 #7021 @sinanozaydin
Hello @kthyng ,

I just did it:
-New version: 1.0.2
-New doi: 10.5281/zenodo.14618844 - https://doi.org/10.5281/zenodo.14618844
