### 2025-01-09 #7021 @kthyng
For future reference, no you don't need to do a new release/version for changes only to the paper. But I'll use what you have done here!
